### 2023-07-30 #599 @paocorrales
agroclimatico:  Índices y Estadísticos Climáticos e Hidrológicos
Nombre de la Persona Encargada: Paola Corrales
Usuario GitHub de la Persona Encargada: @paocorrales
Usuario GitHub de las Otras Autoras del Paquete: @eliocamp, @yabellini, @NatiGattinoni
Repositorio: https://github.com/AgRoMeteorologiaINTA/agroclimatico
Versión Enviada: 1.0.0
Tipo de Envio: Estándar
Editora: @Pakillo
Revisores: @VeruGHub, @pmnatural

Archivo: TBD
Versión Aceptada: TBD
Idioma: es
---

-   Pega el archivo DESCRIPTION completo dentro del siguiente bloque de código.

```
Package: agromet
Title: Índices y Estadísticos Climáticos e Hidrológicos
Version: 1.0.0
    c(person(given = "Yanina",
           family = "Bellini Saibene",
           role = c("ctb"),
           email = "yabellini@gmail.com"),
      person(given = "Elio",
           family = "Campitelli",
           role = c("aut"),
           email = "elio.campitelli@cima.fcen.uba.ar",
      person(given = "Paola",
           family = "Corrales",
           role = c("cre", "aut"),
           email = "paola.corrales@cima.fcen.uba.ar"),
      person(given = "Natalia",
           family = "Gattinoni",
           role = c("aut"),
           email = "gattinoni.natalia@inta.gob.ar"),
      person(family = "INTA",
           role = c("cph")),
      person("Ruida", "Zhong", role = "cph")
      )
Description: Conjunto de funciones para calcular índices y estadísticos climáticos
    hidrológicos a partir de datos tidy. Incluye una función para graficar resultados
    georeferenciados y e información cartográfica.
License: GPL-3 + file LICENSE
Language: es
    scales,
    lmomco,
    sf,
    automap,
    png,
    grid,
    ggnewscale,
    rappdirs,
    kableExtra,
    Rcpp (>= 0.12.0),
    cli,
    SPEI (>= 1.8.1)
    vdiffr,
    covr,
LinkingTo: Rcpp
```

## Alcance

- Por favor, indica qué categoría(s) aplican a este paquete. Las puedes encontrar en nuestras [políticas de inclusión de paquetes (Inglés)](https://ropensci.github.io/dev_guide/policies.html#package-categories). Por favor, tilda todas las apropiadas. Si no estás seguro, te sugerimos que comiences un pre-envío.

	- [ ] recuperación de datos (data retrieval)
	- [ ] extracción de datos (data extraction)
	- [x] munging de datos (data munging)
	- [ ] disposición o declaración de datos (data deposition)
    - [ ] validación y prueba de datos (data validation and testing)
	- [ ] automatización de flujos de trabajo (workflow automation)
	- [ ] control de versiones (version control)
	- [ ] manejo de citas y bibliometría (citation management and bibliometrics)
	- [ ] envoltorios de software científico (scientific software wrappers)
	- [ ] herramientas para trabajo de campo y reproducibilidad (field and lab reproducibility tools)
	- [ ] ligamientos con software de base de datos (database software bindings)
	- [ ] datos geoespaciales (geospatial data)
	- [ ] análisis de texto (text analysis)

- Explica cómo y por qué el paquete encaja dentro de estas categorías (1 a 3 oraciones):
El paquete agromet incluye una serie de funciones para trabajar con datos meteorológicos y calcular distintos índices asociados a aplicaciones agrometeorológicas. Los datos de entrada se organizan usando la filosofía de datos ordenados (o tidy), por lo que las funciones del paquete son genéricas. Pueden aplicarse a cualquier conjunto de datos tabulares independientemente de su origen, orden o nombres de columna.

- ¿Cuál es la audiencia esperada y las aplicaciones científicas de este paquete?
El paquete es ampliamente utilizado por usuaries en distintas dependencias del Instituto Nacional de Tecnología Agropecuaria (INTA)  que brindan servicios e información a los agricultores de Argentina. Sin embargo, les usuaries del paquete se extienden a otras instituciones y países de América Latina ya que casi todas las funciones están desarrolladas para ser de uso general e independiente de los datos recolectados por el INTA y la documentación está escrita integramente en español.

- ¿Hay otros paquetes de R que logren el mismo objetivo? Si los hay, ¿cómo se diferencian del tuyo, o alcanzan nuestro [criterio del mejor de su categoría (documento en Inglés)](https://ropensci.github.io/dev_guide/policies.html#overlap)?

Existen algunos paquetes cómo [frost](https://github.com/anadiedrichs/frost) y [meteor](https://cran.r-project.org/web/packages/meteor/index.html) que incluyen funciones para utilizar datos meteorológicos y calcular indices que en algunos casos están asociadosa la actividad agropecuaria. Otro paquete a mencionar es [agroclim](https://cran.r-project.org/web/packages/agroclim/index.html) que según la documentación también incluye el cálculo de una serie de índices agrometeoroógicos fue archivado por CRAN en abril de este año pero el código esta disponible en un repositorio público.

agromet se diferencia de estos paquetes principalmente por 2 características. Por un lado su documentación está escrita integramente en español lo que facilita su uso en Sudamérica. Por el otro lado, sus funciones estan pensadas para ser compatibles con el ecosistema de paquetes que utilizan la filosofía "tidy", lo que facilita posterior manipulación, visualización y uso de los resultados obtenidos con agromet.

-   (Si aplica) ¿Tu paquete cumple con nuestras [guías de _Ética, Privacidad de Datos e Investigación de Sujetos Humanos_ (documento en Inglés)](https://devguide.ropensci.org/policies.html#ethics-data-privacy-and-human-subjects-research)?
No aplica.

- Si ya has hecho una consulta de pre-envío, por favor pega el enlace al _issue_ correspondiente, una publicación del foro, u otra discusión. Alternativamente, etiqueta al editor (con @tag) con el que te contactaste.
No aplica.

-   (Si aplica) Explique las razones de los [elementos `pkgcheck`](https://docs.ropensci.org/pkgcheck/) que su paquete no puede pasar.

## Revisiones Técnicas

Tilda los siguientes items para confirmar que los has completado:

- [x] He leído la [guía sobre paquetes de rOpenSci](https://devguide.ropensci.org/building.html).
- [x] He leído la [guía para autores](https://devdevguide.netlify.app/authors-guide.html) y pienso mantener este paquete durante al menos 2 años o encontrar un reemplazo.

Este paquete:

- [x] No viola los Términos de Servicio de ningún servicio con los que interactúa.
- [x] Tiene una licencia aceptada por CRAN y OSI.
- [x] Contiene un [archivo README con instrucciones para instalar la versión de desarrollo](https://ropensci.github.io/dev_guide/building.html#readme).
- [x] Incluye [documentación con ejemplos para todas las funciones, creada con roxygen2](https://ropensci.github.io/dev_guide/building.html#documentation).
- [x] Contiene una viñeta (_vignette_) con ejemplos de sus funciones esenciales y su uso.
- [x] Tiene una [suite de tests (documento en Inglés)](https://devguide.ropensci.org/building.html#testing).
- [x] Tiene [integración contínua (documento en Inglés)](https://devguide.ropensci.org/ci.html), incluyendo reporte de cobertura de tests.

## Opciones de Publicación

- [x] ¿Tienes intenciones de subir este paquete a CRAN?
- [ ] ¿Tienes intenciones de enviar este paquete a Bioconductor?

- [ ] ¿Deseas enviar un Artículo de Aplicaciones sobre tu paquete a [Methods in Ecology and Evolution (documento en Inglés)](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/)? Si es así:

<summary>Opciones para MEE</summary>

- [ ] Este paquete es novedoso y será de interés para la mayoría de lectores de la revista.
- [ ] El manuscrito que describe el paquete no tiene más de 3000 palabras y está escrito en Inglés.
- [ ] Tienes intenciones de archivar el código del paquete en un repositorio a largo plazo, que cumple los requerimientos de la revista (mira las [Políticas de Publicación de MEE (documento en Inglés)](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/journal-resources/policy-on-publishing-code.html))
- (*Alcance: Considera los [Objetivos y Alcance de MEE (documento en Inglés)](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/aims-and-scope/read-full-aims-and-scope.html) para tu manuscrito. No otorgamos garatías de que tu manuscrito esté en el ámbito de MEE.*)
- (*Aunque no es requerido, recomendamos tener un manuscrito completamente preparado y en Inglés, al momento de enviar.*)
- (*Por favor, no envíes tu paquete de forma separada a Methods in Ecology and Evolution*)

## Código de Conducta

- [x] Estoy de acuerdo en cumplir el [Código de Conducta de rOpenSci (documento en Inglés)](https://ropensci.github.io/dev_guide/policies.html#code-of-conduct) durante el proceso de revisión, y en mantener mi paquete si éste es aceptado.
