### 2023-09-04 #599 @Pakillo
Gracias @paocorrales y colaboradoras por contribuir su paquete al proceso de revisión de rOpenSci, para el cual actuaré de editor invitado. Cualquier comentario o sugerencia que tengan durante el proceso, por favor no duden en comunicármelo.

Algunos puntos que me gustaría tratar antes de comenzar el proceso de revisión:

## Alcance y solapamiento con otros paquetes

Estoy de acuerdo en la amplia utilidad del paquete. Sólo un comentario: cuando se discute el posible solapamiento con otros paquetes se menciona `agroclim` que "fue archivado por CRAN en abril de este año y el código no parece estar disponible en un repositorio público."

`agroclim` está disponible en este repositorio: https://github.com/rsnotivoli/agroclim. El [motivo](https://cran.r-project.org/web/packages/agroclim/index.html) de expulsión de CRAN parece ser el cambio de email del autor. Parece que la [intención](https://github.com/rsnotivoli/agroclim/issues/1) del autor es restablecer el paquete en CRAN. En cualquier caso, aunque exista solapamiento creo que existe nicho para ambos paquetes. Pero estará bien revisar esta cuestión (sobre todo para evitar posibles duplicidades y reforzar si es posible la funcionalidad existente) durante el proceso de revisión.

Igualmente, como sugerencia más que requisito, pienso que tal vez podría ser útil estudiar la compatibilidad con paquetes como [ClimInd](https://cran.r-project.org/package=ClimInd), que calcula más de 130 índices climáticos diferentes. Por ejemplo, mostrando algunos ejemplos de cómo integrar ambos paquetes para calcular distintos índices.

Por último, veo que existe otro paquete llamado [agrometR](https://github.com/ODES-Chile/agrometR) para descargar información de la Red Agroclimática Nacional de Chile (llamada agromet, https://www.agromet.cl/). ¿Piensan que la similitud en el nombre (agromet / agrometR) y tema (clima) de ambos paquetes podría ser problemático o confuso para futuros usuarios? En tal caso, ¿se plantearían un cambio de nombre del paquete? (ver [guía rOpenSci](https://devguide.ropensci.org/building.html?q=name#package-name-and-metadata)).

## Dependencias

El chequeo automático sugiere que "Some imported packages appear to have no associated function calls; please ensure with author that these 'Imports' are listed appropriately."

Una búsqueda rápida en el código parecen indicar que los paquetes señalados (scales, ggnewscale, kableExtra, Rcpp) sí que se emplean, por lo que creo que podemos ignorar este comentario.

## Nombres de funciones

Hay algunas funciones con nombre duplicado respecto a otros paquetes:

- ith from bioclim
- spei from SPEI
- spi from bayestestR, precintcon, SPEI

Me preocupa particularmente la duplicidad de nombres con el paquete SPEI (funciones `spi` y `spei`). Son funciones para calcular lo mismo, pero tanto los argumentos como el resultado (output de la función) difieren. Pienso que esto puede inducir a confusión a algunos usuarios. No es un requisito pero ¿podría plantearse un ligero cambio de nombre para evitar la confusión? ¿O piensan que mantener el mismo nombre es de hecho recomendable? [Aquí](https://devguide.ropensci.org/building.html?q=function%20nam#function-and-argument-naming) están las directrices de rOpenSci al respecto.

## Documentación

Creo que sería muy útil generar ya una web del paquete para facilitar el proceso de revisión. Esta web puede generarse automáticamente usando `pkgdown::build_site()` (ver [recomendación](https://devguide.ropensci.org/building.html?q=pkgd#website)).

En el README, el tutorial (vignette) y algunas funciones se habla de datos en formato NH y estaciones NH. Estaría bien clarificar a qué se refiere NH.

Por favor añadan unas directrices sobre cómo contribuir (ver https://devguide.ropensci.org/collaboration.html?q=contributi#contributing-guide).

## Chequeo del paquete

Tras ejecutar `devtools::check()` parece que hay algunos problemas menores. Por favor corríjanlos para poder comenzar el proceso de revisión.

Warning: [escalas.R:16] @format requires a value
Warning: [escalas.R:20] @format requires a value
Warning: [escalas.R:24] @format requires a value

## Integración Continua

Por favor modifiquen su GitHub Action para chequear el paquete en la versión actual de R, la anterior, y la que viene (ver [guía](https://devguide.ropensci.org/ci.html#testing-using-different-versions-of-r)). E idealmente tanto en Windows como Mac como Ubuntu. Para ello podría usarse esta [plantilla](https://github.com/r-lib/actions/blob/v2/examples/check-standard.yaml) generada por `usethis::use_github_action("check_standard")`.

## Revisores

Por favor sugieran el nombre de varios revisores potenciales del paquete.

Muchas gracias
