+++ DESCRIPTION
@@ -32,7 +32,7 @@ Language: es
 LazyData: true
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.2.3
-URL: https://github.com/AgRoMeteorologiaINTA/agromet
+URL: https://github.com/AgRoMeteorologiaINTA/agromet, https://agrometeorologiainta.github.io/agromet/
 BugReports: https://github.com/AgRoMeteorologiaINTA/agromet/issues
 Depends: 
     R (>= 2.10)
