+++ _pkgdown.yml
@@ -1,0 +1,4 @@
+url: https://agrometeorologiainta.github.io/agromet/
+template:
+  bootstrap: 5
+lang: es
