+++ pkgdown.yml
@@ -1,0 +1,10 @@
+pandoc: 2.19.2
+pkgdown: 2.0.7
+pkgdown_sha: ~
+articles:
+  estadisticas-e-indices-climaticos: estadisticas-e-indices-climaticos.html
+last_built: 2023-11-01T20:54Z
+urls:
+  reference: https://agrometeorologiainta.github.io/agromet/reference
+  article: https://agrometeorologiainta.github.io/agromet/articles
+
