### 2023-11-01 e0f5667
arregla warning in devtools::check()
