### 2023-11-01 c39f74b
Actualiza github action para chequear con distintas versiones de R
