### 2023-11-30 31d9d3e
Add NEWS.md
