+++ NEWS.md
@@ -1,0 +1,3 @@
+# agroclimr 1.0.0
+
+* Initial CRAN submission.
