+++ README.Rmd
@@ -13,15 +13,15 @@ knitr::opts_chunk$set(
 )
 ```
 
-# agromet <img src='man/figures/logo.png' align="right" height="139" />
+# agroclimr <img src='man/figures/logo.png' align="right" height="139" />
 
 
 <!-- badges: start -->
-[![R build status](https://github.com/AgRoMeteorologiaINTA/agromet/workflows/R-CMD-check/badge.svg)](https://github.com/AgRoMeteorologiaINTA/agromet/actions)
-[![Codecov test coverage](https://codecov.io/gh/AgRoMeteorologiaINTA/agromet/branch/master/graph/badge.svg)](https://codecov.io/gh/AgRoMeteorologiaINTA/agromet?branch=master)
+[![R build status](https://github.com/AgRoMeteorologiaINTA/agroclimr/workflows/R-CMD-check/badge.svg)](https://github.com/AgRoMeteorologiaINTA/agroclimr/actions)
+[![Codecov test coverage](https://codecov.io/gh/AgRoMeteorologiaINTA/agroclimr/branch/master/graph/badge.svg)](https://codecov.io/gh/AgRoMeteorologiaINTA/agroclimr?branch=master)
 <!-- badges: end -->
 
-El paquete {agromet} incluye una serie de funciones para calcular índices y estadísticos climáticos e hidrológicos a partir de datos tidy. Por ejemplo `umbrales()` permite contar la cantidad de observaciones que cumplen una determinada condición y `dias_promedios()` devuelve el primer y último día del año promedio de ocurrencia de un evento. 
+El paquete {agroclimr} incluye una serie de funciones para calcular índices y estadísticos climáticos e hidrológicos a partir de datos tidy. Por ejemplo `umbrales()` permite contar la cantidad de observaciones que cumplen una determinada condición y `dias_promedios()` devuelve el primer y último día del año promedio de ocurrencia de un evento. 
 
 Otras funciones como `spi()` funcionan como wrappers de funciones de otros paquetes y que buscan ser compatibles con el manejo de datos tidy. 
 
@@ -33,7 +33,7 @@ Para instalar la versión de desarrollo desde [GitHub](https://github.com/), us�
 
 ``` r
 # install.packages("remotes")
-remotes::install_github("AgRoMeteorologiaINTA/agromet", build_vignettes = TRUE)
+remotes::install_github("AgRoMeteorologiaINTA/agroclimr", build_vignettes = TRUE)
 ```
 
 
@@ -42,16 +42,16 @@ remotes::install_github("AgRoMeteorologiaINTA/agromet", build_vignettes = TRUE)
 A continuación se muestra el uso de algunas funciones. Podés encontrar más ejemplos y usos en la viñeta, con 
 
 ```r 
-vignette("estadisticas-e-indices-climaticos", "agromet")
+vignette("estadisticas-e-indices-climaticos", "agroclimr")
 ```
 
 Si se quieren utilizar los datos con formato NH se puede utilizar la función `leer_nh()` y opcionalmente acceder a sus metadatos con `metadatos_nh()`.
 
 ```{r message=FALSE, warning=FALSE}
-library(agromet)
+library(agroclimr)
 library(dplyr)
 
-archivo <- system.file("extdata", "NH0358.DAT", package = "agromet")
+archivo <- system.file("extdata", "NH0358.DAT", package = "agroclimr")
 
 datos <- leer_nh(archivo)
 ```
@@ -88,4 +88,4 @@ datos_aleatorios %>%
 
 ## Cómo contribuir
 
[…]
