+++ DESCRIPTION
@@ -1,4 +1,4 @@
-Package: agromet
+Package: agroclimr
 Title: Índices y Estadísticos Climáticos e Hidrológicos
 Version: 1.0.0
 Authors@R: 
@@ -32,8 +32,8 @@ Language: es
 LazyData: true
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.2.3
-URL: https://github.com/AgRoMeteorologiaINTA/agromet, https://agrometeorologiainta.github.io/agromet/
-BugReports: https://github.com/AgRoMeteorologiaINTA/agromet/issues
+URL: https://github.com/AgRoMeteorologiaINTA/agroclimr, https://agrometeorologiainta.github.io/agroclimr/
+BugReports: https://github.com/AgRoMeteorologiaINTA/agroclimr/issues
 Depends: 
     R (>= 2.10)
 Imports: 
