+++ R/RcppExports.R
@@ -2,6 +2,6 @@
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
 C_pdsi <- function(P, PE, AWC, s_yr, e_yr, calib_s_yr, calib_e_yr, sc, K1_1, K1_2, K1_3, K2, p, q) {
-    .Call('_agromet_C_pdsi', PACKAGE = 'agromet', P, PE, AWC, s_yr, e_yr, calib_s_yr, calib_e_yr, sc, K1_1, K1_2, K1_3, K2, p, q)
+    .Call('_agroclimr_C_pdsi', PACKAGE = 'agroclimr', P, PE, AWC, s_yr, e_yr, calib_s_yr, calib_e_yr, sc, K1_1, K1_2, K1_3, K2, p, q)
 }
 
