### 2023-12-01 528da25
actualizo ejemplos y viñeta
