+++ README.Rmd
@@ -45,7 +45,7 @@ A continuación se muestra el uso de algunas funciones. Podés encontrar más ej
 vignette("estadisticas-e-indices-climaticos", "agroclimr")
 ```
 
-Si se quieren utilizar los datos con formato NH se puede utilizar la función `leer_nh()` y opcionalmente acceder a sus metadatos con `metadatos_nh()`.
+Si se quieren utilizar los datos con formato NH (archivos con columnas de ancho fijo usado por el INTA para distribuir los datos de las estaciones meteorológicas de su red) se puede utilizar la función `leer_nh()` y opcionalmente acceder a sus metadatos con `metadatos_nh()`.
 
 ```{r message=FALSE, warning=FALSE}
 library(agroclimr)
