### 2023-12-01 f3aac6d
Actualiza test de spi
