+++ NEWS.md
@@ -1,3 +1,4 @@
 # agroclimr 1.0.0 (30/11/2023)
 
 * Cambia el nombre del paquete de agromet a agroclimr
+* Cambia el nombre de las funciones `spi()` y `spei()` a `spi_indice()` y `spei_indice()`
