+++ man/spi_indice.Rd
@@ -1,12 +1,12 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/spi.R
-\name{spi}
-\alias{spi}
-\alias{spei}
+\name{spi_indice}
+\alias{spi_indice}
+\alias{spei_indice}
 \alias{spi_referencia}
 \title{Calcula el SPI}
 \usage{
-spi(
+spi_indice(
   fecha,
   precipitacion,
   escalas,
@@ -15,7 +15,7 @@ spi(
   ...
 )
 
-spei(fecha, balance, escalas, distribucion = "log-Logistic", ...)
+spei_indice(fecha, balance, escalas, distribucion = "log-Logistic", ...)
 
 spi_referencia(fecha, precipitacion)
 }
@@ -50,8 +50,8 @@ Un data.frame con columnas
 }
 }
 \description{
-Calcula esl Índice Estandarizado de Precipitación para distintas escalas. El \code{spi} toma valores
-de precipitación mientras que el \code{spei} toma valores del balance entre precipitación y evapotranspiración
+Calcula esl Índice Estandarizado de Precipitación para distintas escalas. El \code{spi_indice} toma valores
+de precipitación mientras que el \code{spei_indice} toma valores del balance entre precipitación y evapotranspiración
 potencial. Internamente hacen lo mismo; la única diferencia es la distribución teórica
 usada por defecto para ajustar los datos.
 }
@@ -61,7 +61,7 @@ datos <- data.frame(fecha = seq(as.Date("1985-01-01"), as.Date("2015-12-01"), by
 set.seed(42)
 datos$pp <- rgamma(nrow(datos), shape = 2, scale = 10)
 
-with(datos, spi(fecha, pp, escalas = 1:5))
+with(datos, spi_indice(fecha, pp, escalas = 1:5))
 
 # Si entran nuevos datos y hay que calcular el spi nuevamente pero sin que
 # cambien los valores viejos, hay que usar `referencia`
@@ -72,11 +72,11 @@ nuevos_datos$pp <- rgamma(nrow(nuevos_datos), shape = 2, scale = 10)
 nuevos_datos <- rbind(datos, nuevos_datos)
 
 # Usando un vector lógico
-with(nuevos_datos, spi(fecha, pp, escalas = 1:5,
+with(nuevos_datos, spi_indice(fecha, pp, escalas = 1:5,
                        referencia = data.table::year(fecha) < 2016))
 
 # O un data.frame
-with(nuevos_datos, spi(fecha, pp, escalas = 1:5,
+with(nuevos_datos, spi_indice(fecha, pp, escalas = 1:5,
                        referencia = spi_referencia(datos$fecha, datos$pp)))
 
 
