+++ NAMESPACE
@@ -23,8 +23,8 @@ export(pdsi_ac)
 export(pdsi_coeficientes)
 export(scale_color_inta)
 export(scale_fill_inta)
-export(spei)
-export(spi)
+export(spei_indice)
+export(spi_indice)
 export(spi_referencia)
 export(theme_inta_mapa)
 export(umbrales)
