+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
   estadisticas-e-indices-climaticos: estadisticas-e-indices-climaticos.html
-last_built: 2023-11-30T22:17Z
+last_built: 2023-12-01T15:50Z
 urls:
   reference: https://agrometeorologiainta.github.io/agromet/reference
   article: https://agrometeorologiainta.github.io/agromet/articles
