+++ tests/testthat/test-leer_surfer.R
@@ -1,5 +1,5 @@
 test_that("multiplication works", {
-  escala <- system.file("extdata", "escala_pp_mensual.lvl", package = "agroclimr")
+  escala <- system.file("extdata", "escala_pp_mensual.lvl", package = "agroclimatico")
 
   expect_known_output(leer_surfer(escala), "escala_surfer")
 })
