+++ tests/testthat/test-mapas.R
@@ -2,7 +2,7 @@
 skip_if_offline()
 
 # Para testear también que descargue el mapa correctamente
-agroclimr:::borrar_cache()
+agroclimatico:::borrar_cache()
 
 test_mapa <- function(mapa_fun) {
 
