+++ man/agroclimatico-package.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/agroclimr-package.R
+% Please edit documentation in R/agroclimatico-package.R
 \docType{package}
-\name{agroclimr-package}
-\alias{agroclimr}
-\alias{agroclimr-package}
-\title{agroclimr: Índices y Estadísticos Climáticos e Hidrológicos}
+\name{agroclimatico-package}
+\alias{agroclimatico}
+\alias{agroclimatico-package}
+\title{agroclimatico: Índices y Estadísticos Climáticos e Hidrológicos}
 \description{
 \if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}
 
@@ -13,9 +13,9 @@ Conjunto de funciones para calcular índices y estadísticos climáticos hidrol�
 \seealso{
 Useful links:
 \itemize{
-  \item \url{https://github.com/AgRoMeteorologiaINTA/agroclimr}
-  \item \url{https://agrometeorologiainta.github.io/agroclimr/}
-  \item Report bugs at \url{https://github.com/AgRoMeteorologiaINTA/agroclimr/issues}
+  \item \url{https://github.com/AgRoMeteorologiaINTA/agroclimatico}
+  \item \url{https://agrometeorologiainta.github.io/agroclimatico/}
+  \item Report bugs at \url{https://github.com/AgRoMeteorologiaINTA/agroclimatico/issues}
 }
 
 }
