+++ NAMESPACE
@@ -34,4 +34,4 @@ importFrom(data.table,":=")
 importFrom(data.table,.BY)
 importFrom(magrittr,"%>%")
 importFrom(stats,ts)
-useDynLib(agroclimr)
+useDynLib(agroclimatico)
