### 2023-12-18 #599 @Pakillo
Muchas gracias por la respuesta y todas las mejoras al paquete @paocorrales et al. Yo creo que está casi listo para pasar a revisión.

Solo tres pequeños detalles por ahora:

- El fichero `ejemplos.Rmd` (y su resultado `ejemplos.md`), ambos en el directorio raíz, parecen obsoletos. ¿Se pueden eliminar?

- El ejemplo con `dias_promedio` en el README (https://github.com/AgRoMeteorologiaINTA/agroclimatico?tab=readme-ov-file#d%C3%ADas-promedio) da un warning debido a los últimos cambios en dplyr. ¿Podrían sustituir `summarise` por `reframe` como sugiere el warning?

- El chequeo automático avisa también de que hay [una línea](https://github.com/AgRoMeteorologiaINTA/agroclimatico/blob/2d86dbe80bf547ed67fe6ae19204bd42d3db32e2/R/scpdsi-internal.R#L26) con punto y coma al final. Esto no tiene ninguna importancia ni afecta al funcionamiento, pero sugiere su eliminación.

Procedo a continuación a buscar revisores. Si pueden arreglar estos pequeños detalles mientras tanto, estupendo.
