### 2023-12-19 e20a6b1
Elimina ejemplos repetidos
