+++ ejemplos.Rmd
@@ -1,157 +1,0 @@
----
-title: "Ejemplos de uso del paquete {agroclimatico}"
-output: github_document
----
-
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = TRUE)
-
-# remotes::install_github("AgRoMeteorologiaINTA/agroclimatico")
-library(agroclimatico)
-library(dplyr)
-library(lubridate)
-library(ggplot2)
-```
-
-## Lectura de archivo NH
-
-Asumiendo que los datos de estaciones de la base de datos NH están en una carpeta "datos" dentro del proyecto, la función `leer_nh()` funciona así:
-
-```{r lectura}
-datos <- leer_nh("datos/NH0011.DAT")
-
-head(datos)
-```
-
-Además la funciónn `metadatos_nh()` devuelve un data frame con los metadatos de las estaciones que se puede utilizar para conocer el nombre y ubicación de las estaciones. 
-
-```{r}
-metadatos_nh(codigo = "0011")
-```
-
-
-## Conteo de días que cumplen con alguna condición
-
-Para contar la cantidad de diás con temperatura por encima de un umbral o días con precipitación igual a cero (o básicamente que cualquier variable cumpla con cualquier condición) se utiliza la función `umbrales()` [o inserte aquí un mejor nombre].
-
-Por ejemplo si queremos la cantidad de días donde la temperatura máxima fue mayor o igual a 30 grados en toda la serie.
-
-```{r extremos-1}
-datos %>% 
-  group_by(fecha = lubridate::floor_date(fecha, "1 month")) %>% 
-  summarise(umbrales(t_30 = t_max >= 30, 
-                     secos = precip <= 0.1,
-                     t_0  = t_min <= 0,
-                     t_15 = t_max >= 15)) %>% 
-  filter(lubridate::month(fecha) == 1) %>% 
-  ggplot(aes(fecha, prop)) +
-  geom_line(aes(color = extremo))
-```
-
-Pero si quisiéramos hacer el mismo conteo para cada año:
-
-```{r extremos-2}
-datos %>% 
-  group_by(anio = floor_date(fecha, "year")) %>% 
-  summarise(extremos(t_30 = t_max >= 30))
-```
-
-Esta función también acepta contar cantidad de días según distintos umbrales. De yapa, un gráfico.
-
-```{r extremos-3}
-datos %>% 
-  group_by(anio = floor_date(fecha, "year")) %>% 
-  summarise(extremos(t_30 = t_max >= 30,
-                     t_35 = t_max >= 35)) %>% 
-  ggplot(aes(anio, N)) +
-  geom_line(aes(color = extremo))
-```
-
-## SPI
-
-La función `spi_indice()` calcula el SPI para distintas escalas:
-
-```{r}
-datos %>% 
-  summarise(spi_indice(fecha, prcp, 1:24)) %>% 
-  head()
-```
-
-La salida se puede graficar fácilmente con ggplot
-
-```{r}
-referencia <- datos %>% 
-  filter(year(fecha) <= 1980) %>% 
-  summarise(spi_referencia(fecha, prcp))
-
-datos %>% 
-  summarise(spi_indice(fecha, prcp, 1:24, 
-                referencia = referencia)) %>% 
-  ggplot(aes(fecha, escala)) +
-  geom_contour_filled(aes(z = spi, fill = stat(level_mid))) +
-  scale_fill_gradient2()
-```
-
[…]
