+++ ejemplos.md
@@ -1,158 +1,0 @@
-Ejemplos de uso del paquete {agromet}
-================
-
-## Lectura de archivo NH
-
-Asumiendo que los datos de estaciones de la base de datos NH están en
-una carpeta “datos” dentro del proyecto, la función `leer_nh()` funciona
-así:
-
-``` r
-datos <- leer_nh("datos/NH0011.DAT")
-```
-
-    ## Warning: 1 parsing failure.
-    ##   row col expected actual               file
-    ## 20969 etp  4 chars      3 'datos/NH0011.DAT'
-
-``` r
-head(datos)
-```
-
-    ##    codigo codigo_nh      fecha t_max t_min prcp lluvia_datos lluvia llovizna
-    ## 1:      5      0011 1956-04-01  25.0  11.2  0.0            0     NA       NA
-    ## 2:      5      0011 1956-04-02  28.7  16.5  2.4            1     NA       NA
-    ## 3:      5      0011 1956-04-03  28.7  20.7  1.5            1     NA       NA
-    ## 4:      5      0011 1956-04-04  26.7  19.8  0.0            0     NA       NA
-    ## 5:      5      0011 1956-04-05  25.6  15.6  0.0            0     NA       NA
-    ## 6:      5      0011 1956-04-06  27.2  16.4  0.9            1     NA       NA
-    ##    granizo nieve t_aire_max t_aire_min t_suelo_max t_suelo_min heliofania
-    ## 1:      NA    NA         NA         NA          NA          NA       11.1
-    ## 2:      NA    NA         NA         NA          NA          NA        2.2
-    ## 3:      NA    NA         NA         NA          NA          NA        3.8
-    ## 4:      NA    NA         NA         NA          NA          NA        0.0
-    ## 5:      NA    NA         NA         NA          NA          NA        3.0
-    ## 6:      NA    NA         NA         NA          NA          NA        0.4
-    ##    heliofania_rel p_vapor hr td rocio viento_med viento_2m  rad etp
-    ## 1:             94      NA NA NA    NA         NA        NA 22.6  NA
-    ## 2:             18      NA NA NA    NA         NA        NA  9.0  NA
-    ## 3:             32      NA NA NA    NA         NA        NA 11.4  NA
-    ## 4:              0      NA NA NA    NA         NA        NA  5.7  NA
-    ## 5:             25      NA NA NA    NA         NA        NA 10.0  NA
-    ## 6:              3      NA NA NA    NA         NA        NA  6.2  NA
-
-Además la funciónn `metadatos_nh()` devuelve un data frame con los
-metadatos de las estaciones que se puede utilizar para conocer el nombre
-y ubicación de las estaciones.
-
-## Conteo de días que cumplen con alguna condición
-
-Para contar la cantidad de diás con temperatura por encima de un umbral
-o días con precipitación igual a cero (o básicamente que cualquier
-variable cumpla con cualquier condición) se utiliza la función
-`extremos()` \[o inserte aquí un mejor nombre\].
-
-Por ejemplo si queremos la cantidad de días donde la temperatura máxima
-fue mayor o igual a 30 grados en toda la serie.
-
-``` r
-datos %>% 
-  summarise(extremos(t_30 = t_max >= 30))
-```
-
-    ##       N      prop  na extremo
-    ## 1 12616 0.5532605 721    t_30
-
[…]
