### 2023-12-19 #599 @paocorrales
Muchas gracias @Pakillo, resolví todos los detalles mencionados :smiley:
