### 2024-01-19 #599 @Pakillo
Hola @paocorrales. Disculpas por la demora, ha costado bastante encontrar un segundo revisor. Ya está todo en marcha.

@VeruGHub ¿cómo va la revisión? Espero que pueda estar lista pronto, para que los autores puedan empezar a revisar las sugerencias. Si puedo ayudar en algo no dude en avisarme.

¡Gracias!
