### 2024-01-22 #599 @VeruGHub
@Pakillo estoy con la revisión hoy. Disculpad la demora
