### 2024-02-09 #599 @VeruGHub
## Revisión de un paquete

No tengo ninguna relación de trabajo con los autores

-   [x] Como revisor confirmo que no tengo [conflictos de interés](https://devguide.ropensci.org/policies.html#coi) para poder hacer la revisión de este trabajo (si no estas segura si tienes un conflicto por favor entra en contacto con tu editor *antes* de arrancar con la revisión.

#### Documentación

El paquete incluye todos los siguiente tipos de documentación:

-   [x] **Una declaración de necesidades** que claramente describe las necesidades que el software esta diseñado a resolver y el public meta que busca atender en el archivo README
-   [x] **Instrucciones de instalación** de la versión en desarrollo del paquete incluyendo cualquier dependencia no-estándar en el archivo README
-   [ ] **Viñeta(s)** demostrando la funcionalidad principal que ademas corren localmente
-   [x] **Documentación de las funciones** exportadas
-   [ ] **Ejemplos** (que corren localmente) para todas las funciones exportadas
-   [x] **Directrices comunitarias** incluyendo una guia de contribución en el archivo README o el archivo CONTRIBUTING y un archivo DESCRIPTION que incluye `URL`, `BugReports` and `Maintainer` (todas en inglés por concenvión y para que puedan ser autogeneradas con `Authors@R`).

#### Funcionalidad

-   [x] **Instalación:** La instalación se completa con éxito tal como fue documentada.
-   [x] **Funcionalidad:** Toda afirmación de funcionalidad del software se confirma como existente.
-   [ ] **Desempeño:** Toda afirmación de desempeño del software se confirma como alcanzada.
-   [ ] **Pruebas automáticas:** Hay pruebas unitarias que cubren las funciones esenciales dentro del paquete con un rango razonable de entradas y condiciones. Todas las pruebas corren en la maquina local.
-   [ ] **Directrices de empaque**: El paquete cumple con las directrices de empaque de rOpenSci.

Estimación de horas dedicadas a la revisión: 8

-   [x] Si la o las persona(s) autora(s) lo considera(n) apropiado, yo estoy de acuerdo con que me reconozcan como revisor del paquete (el rol "rev') en la el archivo DESCRIPTION del paquete.

### Comentarios de la revisión

En primer lugar, quiero agradecer la oportunidad de revisar este paquete y espero que los comentarios sirvan para mejorar en los puntos que los autores consideren oportunos. Creo que los autores han hecho un gran trabajo compilando las funciones relacionadas con la gestión de datos climáticos que son útiles para el INTA.

He revisado el paquete desde el punto de vista de un usuario y tengo algunos comentarios generales y también específicos de cada función. Incluyo en la revisión código que he ido ejecutando y con comentarios en partes que no han funcionado correctamente y reflexiones.

Creo que algunos aspectos formales de la documentación pueden ser mejorados. Para empezar, la funcionalidad del paquete no está completamente definida en la documentación (Readme). El paquete calcula índices y estadísticos climáticos e hidrológicos a partir de datos tidy, pero también permite representar esta información y otra en mapas en Argentina, y permite leer datos climáticos en el formato del INTA. Creo que toda esta información debería de estar incluida en el Readme. En varios puntos de la documentación de las funciones (ver debajo) he detectado que no se especifica el formato concreto en el que los argumentos deben declararse y también argumentos cuya descripción es confusa. Mejorar estas descripciones e incluir programación defensiva relativa a los argumentos (en general faltante) ayudaría mucho a los usuarios.

Creo que la funcionalidad de spi_indice y spei_indice está ya contenida en el paquete SPEI. A menos que se aclare cual es la novedad de las funciones escritas para agroclimático, sugiero eliminar estas funciones. Además, no he conseguido correr las funciones scale_color/fill_inta y pdsi/pdsi_ac (ver debajo los comentarios).

Finalmente, en el logotipo del paquete pone "AgroMET" lo que es un poco raro si el paquete se llama "agroclimatico".

library(agroclimatico)

vignette("estadisticas-e-indices-climaticos", "agroclimatico") #no me funciona
archivo <- system.file("extdata", "NH0358.DAT", package = "agroclimatico")
datos <- leer_nh(archivo)
data(NH0358)
```

#### dias_promedio

No calcula ningun estadístico meteorológico como dice la descripción si no que extrae dia, mes y dia juliano de un set de datos. Previamente el usuario puede haber extraido los días que representan un evento meteorológico de interés, pero no tiene por qué.

La descripción de la funcionalidad no es clara en la ayuda. Sugiero poner "Calcula el valor promedio anual del primer y ultimo dia para una serie de fechas dada". En los ejemplos habría que cambiar "summarise" por "reframe" como aparece en el Readme. En "Value" se dice que el output tiene "variables extras en el caso de hacer el cálculo para distintos grupos", pero ¿cómo se puede especificar para qué grupos quieres el cálculo?

Erratas en la documentación: "seria de fechas"

dias_promedio(datos$fecha) #Ok
dias_promedio(NH0358$fecha) #Ok

helada <- NH0358 |>
  filter(t_min <= 0)

dias_promedio(helada$fecha) #Ok

helada |> summarise(dias_promedio(fecha)) #En un pipe, solo si la funcion se escribe dentro de un summarise o reframe se consigue la funcionalidad descrita

#La función no devuelve un output que encaje bien con la filosofía de tidyverse y es difícil de utilizar en un pipe:

helada |> dias_promedio(fechas = fecha)
helada |> mutate(diasprom = dias_promedio(fecha))

```

#### completar_serie

No se especifica el formato del vector fecha ni de rango en la documentación.

#Ejecuto el ejemplo #Ok
ejemplo <- data.frame(fechas = seq(as.Date("1985-01-01"), as.Date("2015-12-01"),
                             by = "1 month"),
                    pp = 1)

datos_perdidos <- sample(nrow(ejemplo), nrow(ejemplo)/10)
ejemplo$pp[datos_perdidos] <- NA
datos_incompletos <- na.omit(ejemplo)

completar_serie(datos_incompletos, fechas, resolucion = "1 mes")

#No funciona, pero sigue las instrucciones de la documentación:
completar_serie(datos = datos,
                fecha = fecha,
                resolucion = "día",
                rango = as.Date(c("1951-01-01", "1952-12-31")))

```

#### decil y anomalia_porcentual

En la descripción se necesita incluir una definición más técnica de qué valores se devuelven, es decir, que hace la función decil y la anomalia_porcentual y referencias. En estadística, los deciles de una serie de datos son estadísticos descriptivos que representan los valores que limitan la serie de datos en 10 partes. Al aplicar la función sin tener mucho detalle en la descripción yo hubiera esperado obtener eso. Sin embargo, en el ejemplo ejecutado debajo obtengo un vector de igual longitud de la serie de datos. Sugiero mejorar la descripción en este sentido. La anomalía porcentual ¿respecto a qué valor de la serie de referencia se calcula? Debería especificarse. ¿Por qué para calcular la anomalia es necesario especificar que no se consideren los NA, pero para decil no? Parece una cuestión de inconsistencia en cómo están escritas las funciones.

decil(datos$t_max) #Ok

anomalia_porcentual(datos$t_max, na.rm = TRUE) #Ok
```

#### scala_temp/pp_XXX

La documentación está incompleta. No hay ejemplos de uso ni se entiende bien que hacen estas funciones. Tampoco se especifica qué valores retorna.

#### ith

Funciona correctamente y está bien documentada.

NH0358 <- NH0358 %>% #Ok
  mutate(t_media = (t_max + t_min)/2) %>%
  mutate(ith = ith(t_media, hr))

hist(NH0358$ith)
```

#### metadatos_nh

Para facilitar la comprensión a los usuarios convendría especificar que son "estaciones NH". La descripción es incompleta en el sentido de que sólo se especifica una parte de la información que devuelve la función.

Erratas en la documentación: "provincia o provincia", "especificar" debería ser "específicas".

metadatos_nh(codigo = c("0001", "0011")) #Ok
metadatos_nh(provincia ="Chaco") #Ok
metadatos_nh(organismo = "SMN") #Ok
metadatos_nh(lat = c(-30, -20), lon = c(-65, -55)) #Ok
```

#### kable_inta

Sustituir "Setear" por "Formatear" o similar. Funciona correctamente.

kable_inta(datos[1:10,1:5]) #Ok
kable_inta(metadatos_nh(provincia ="Chaco")) #Ok
```

#### leer_surfer

En el ejemplo se sobreescribe un objeto con el nombre de otra función implementada en el paquete (scala_temp/pp_XXX). Debería estar explicado en algún sitio cual es el link entre ambas.

Erratas en la documentación: "laeer".

escala <- system.file("extdata", "escala_pp_mensual.lvl", package = "agroclimatico")

escala_prueba <- leer_surfer(escala) #Ok
scales::show_col(escala_prueba$colores)
```

#### mapa_XXX

Añadiría en la descripción los posibles valores para el argumento "provincias".

ggplot() +
  geom_sf(data = mapa_argentina()) #Ok

ggplot() +
  geom_sf(data = mapa_provincias()) #Ok

ggplot() +
  geom_sf(data = mapa_provincias(provincias = "Chaco", departamentos =  TRUE)) #Ok

ggplot() +
  geom_sf(data = mapa_provincias(provincias = "Rio Negro", departamentos = FALSE)) #Ok

ggplot() +
  geom_sf(data = mapa_argentina_limitrofes()) #Ok

ggplot() +
  geom_sf(data = mapa_departamentos(provincias = "Chaco")) #Ok
```

#### mapear

Erratas en la documentación: "veriable", "a el", "masyores", "1500m".

El argumento "breaks" no está bien definido. Habría que especificar que se refiere a que hay que introducir valores (¿uno o varios?) en el rango del argumento "valor".

No se explican las funciones coord_argentina ni theme_inta_mapa en ningún sitio. ¿Se utilizan sólo internamente? En ese caso no deberían estar exportadas.

prueba <- metadatos_nh()

mapear(prueba$altura, #Ok
       prueba$lon,
       prueba$lat,
       breaks = c(0, 100, 400, 1000, 1500),
       cordillera = 1400,
       variable = "mm")

```

#### olas y umbrales

No se explica en la documentación como hay que especificar los umbrales. Es poco habitual que se pueda poner cualquier palabra como nombre de un argumento y una mala praxis porque se puede sobreescribir algo importante en la función. Sería más adecuado que hubiera un argumento para los umbrales y que éstos se pudieran introducir en forma de lista con nombres (ej. umbral = list(calor = "t_max \> 20"))

Corregir algunas erratas en la documentación de olas: "mbral", "drtn"

Corregir algunas erratas en la documentación de umbrales: "``` ⁠N`` (numérico) ocurrencia del evento *⁠ ```prop' "

olas(datos$fecha, heatwave = datos$t_max > 20) #Ok

umbrales(heatwave = datos$t_max > 20) #Ok

```

#### pdsi/pdsi_ac

No funciona indicandole los argumentos no definidos por defecto.

Habría que incluir en la documentación las referencias que se citan e incluir las unidades de la precipitacion y la etp. Es redundante incluir un argumento para los coeficientes en forma de lista y luego argumentos para cada coeficiente por separado. Sugiero eliminar una de las dos formas.

Corregir algunas erratas de la ayuda: "list de coeficientes", "El el cálculo".

pdsi(precipitacion = c(500, 600), #No funciona
     etp = c(1000, 1200))

pdsi_ac(precipitacion = c(500, 600), #No funciona
     etp = c(1000, 1200))
```

#### scale_color/fill_inta pdsi/pdsi_ac

El ejemplo no funciona. Tampoco funciona otro ejemplo que he creado siguiendo las instrucciones de la documentación.

Corregir erratas en la documentación: "escala.)", "tiene prioridad por sobre", "deinidos"

prueba <- metadatos_nh()
prueba$altura <- as.numeric(prueba$altura)

ggplot(prueba) + #No funciona
  geom_point(aes(x = lon, y = lat, color = altura)) +
  scale_color_inta(escala = escala_pp_diaria)

set.seed(496) #El ejemplo no funciona
datos_aleatorios <- subset(metadatos_nh())
datos_aleatorios <- data.frame(datos_aleatorios,
pp = rgamma(nrow(datos_aleatorios), 0.5, scale = 1)*70)

ggplot(datos_aleatorios, aes(lon, lat)) +
geom_contour(aes(z = pp, fill = after_stat(level))) +
scale_fill_inta(escala = escala_pp_diaria)
```

#### spei_indice/spi_indice

La funcionalidad está repetida respecto a las funciones spi y spei del paquete SPEI. A menos que se aclare qué aporta esta función sugiero eliminarla y crear una viñeta que incluya cómo se usa el paquete SPEI con datos en el formato del INTA. Si he entendido bien, las fechas deben introducirse como meses ¿no? Por favor, especificar esto si se mantiene la función. En la documentación no se entiende bien que es el argumento "escalas" y no se cita el paquete SPEI.

prueba <- data.frame(fecha = seq(as.Date("1985-01-01"), as.Date("2015-12-01"), by = "1 month"))
prueba$pp <- rgamma(nrow(prueba), shape = 2, scale = 10)

spi_indice(prueba$fecha, prueba$pp, escalas = 1)

#Misma funcionalidad que en paquete SPEI:
spi_agro <- spi_indice(fecha = prueba$fecha, precipitacion = prueba$pp, escalas = 1)

spi_spei <- SPEI::spi(data = prueba$pp, scale = 1)

identical(spi_agro$spi, as.vector(spi_spei$fitted))

```
