### 2024-02-12 #599 @Pakillo
¡Muchas gracias @VeruGHub por su detallada revisión! Seguro que será muy útil.
