### 2024-02-13 #599 @Pakillo
Estimada @paocorrales y coautores:

Ya tenemos las revisiones de @pmnatural y @VeruGHub, a quienes agradezco su esfuerzo. Espero que sus comentarios sean útiles. Añado algunos comentarios al respecto:

## README

Ambas revisoras comentan que les gustaría ver más información sobre el paquete ahí. Yo la veo correcta (sobre todo al contar con un magnífico tutorial complementario (vignette)), pero tal vez se pueda añadir algo más o re-estructurar la información del README para que quede más claro, al menos en cuanto a los grandes bloques de funcionalidad del paquete (lectura de datos INTA, indicadores, mapeo, etc). Una sugerencia para que quede más clara la funcionalidad global del paquete es agrupar las funciones en el [índice](https://agrometeorologiainta.github.io/agroclimatico/reference/index.html) por temáticas (ver https://r-pkgs.org/website.html#index-organization).

En el primer párrafo del Readme hay una errata (`dias_promedios()` cuando la función se llama `dias_promedio()`). También sugiero poner 'tidy' entre comillas, y tal vez enlazar esa palabra con algún documento que explique qué significa tidy data para los que no estén familiarizados con el concepto (por ejemplo, https://es.r4ds.hadley.nz/12-tidy.html en español, o https://tidyr.tidyverse.org/articles/tidy-data.html o http://www.jstatsoft.org/v59/i10/paper en inglés).

En el segundo párrafo del Readme, la función `spi` se llama ahora `spi_indice`.

En el tercer párrafo se menciona por primera vez el INTA. Es posible que muchos no conozcan a qué se refiere el acrónimo, por lo que añadiría el nombre completo del instituto y un enlace a la web oficial.

Es cierto que el logo refleja el antiguo nombre del paquete (agromet), ¿tal vez sea buena idea modificarlo?

La 'vignette' a mí si me funciona, siempre que se especifique `build_vignettes = TRUE` al instalar, como explica el README. Por cierto, gran tutorial mostrando las distintas funciones del paquete.

## dias_promedio

En esta y otras funciones, como indica @VeruGHub, sería conveniente reemplazar `summarise` por `reframe` en todos los ejemplos donde aparece tal warning. Coincido también en que no está claro cómo utilizar esta función con datos agrupados (dado que el único argumento de la función es un vector de fechas).

## decil

En la función decil, pienso que sería útil dar la opción de devolver tanto el número del decil (1, 2, 3, ...) como el valor numérico (ej. 7.7 correspondente al decil 1, etc). Por ejemplo viendo el ejemplo en https://agrometeorologiainta.github.io/agroclimatico/articles/estadisticas-e-indices-climaticos.html#deciles no me queda claro cómo interpretar en qué decil nos encontramos simplemente viendo el valor de la columna 'decil'.

## mapear

Esta función me parece muy útil pero me pregunto si podría rediseñarse para que el primer argumento sea un data.frame o tibble, siguiendo la costumbre en tidyverse. Esto permitiría llamarla directamente dentro de un pipe, p. ej.:

```r
datos_aleatorios |>
  mapear(pp, lon, lat,
       cordillera = TRUE,
       escala = escala_pp_diaria,
       variable = "mm",
       titulo = "Precipitación aleatoria",
       fuente = "Fuente: datos de ejemplo")
```

## olas y umbrales

Estas funciones también me gustan mucho y me parecen muy útiles, pero igualmente me pregunto (a modo de reflexión) si el primer argumento podría ser un data.frame o tibble, dado que se usan varias variables del mismo para calcular las olas. Esto podría permitir además que no fuese necesario usar summarise/reframe, p. ej:

```r
olas(NH0358, calor = t_max > 20, frio = t_min <= 0)
```

## SPI, SPEI

También aquí me planteo si no sería conveniente (siguiendo la filosofía tidyverse) que el primer argumento fuera siempre un data.frame o tibble. De ese modo no habría que usar `with`, como en el ejemplo.

Estoy de acuerdo con @VeruGHub en que sería bueno explicar la utilidad de estas funciones y en qué difiere esta implementación de la del paquete SPEI (utilizado internamente).

## General

En los ejemplos sería conveniente simplificar los datasets al máximo, dejando solamente lo necesario para mostrar lo que hace la función. Por ejemplo, en la función `decil` se usa el dataset `NH0358` completo, y el objeto generado es muy grande (filas y columnas) y difícil de leer, hasta el punto de que cuesta ver el resultado de aplicar la función. Sería recomendable reducir el número de filas y columnas al mínimo imprescindible (o mostrar solo una selección de filas cuando ello sea suficiente, ej. en `metadatos_nh`). También se aplica a los ejemplos de otras funciones como `completar_serie`, `ith`, `olas`, `pi_indice`, etc.

## Erratas

Hay algunas erratas en la documentación. Podría usarse el paquete [spelling](https://docs.ropensci.org/spelling/) o herramientas similares para facilitar su detección y corrección.

Nada más por ahora. Espero que mis comentarios, unidos a los de las revisoras, sean útiles para revisar el paquete. Por favor no duden en comentar cualquier duda o discrepancia que puedan tener. Muchos de los cambios propuestos son sugerencias o invitaciones a la reflexión. Quedo pues a la espera de su revisión.

Saludos cordiales

Paco
