### 2024-05-17 #599 @Pakillo
Estimadas @paocorrales, @eliocamp, @yabellini, @NatiGattinoni:

Espero que vaya todo bien. Me preguntaba cómo iba avanzando la revisión del paquete. Si puedo ayudar en algún aspecto o tienen cualquier duda o comentario no duden en decirlo.

Gracias
