### 2024-05-27 bf3cab7
Agrego datos mensuales de estaciones NH para 2019
