### 2024-05-27 18df2fd
Mejora documentación de escalas de colores y funciones asociadas
Close #45
