### 2024-05-27 4014160
Actualizo hex. Gracias @yabellini!
Close #40
