### 2024-05-27 d9473d9
Agrego secciones al indice de funciones en la web
