### 2024-05-27 81244f9
Mejora documetación de tabla
