### 2024-05-27 ab03b51
Arregla typos
Close #46
