+++ man/metadatos_nh.Rd
@@ -13,7 +13,7 @@ metadatos_nh(
 )
 }
 \arguments{
-\item{codigo, provincia, organismo}{carácter o vector de caracteres para filtrar según código de estación, provincia o provincia.}
+\item{codigo, provincia, organismo}{carácter o vector de caracteres para filtrar según código de estación, provincia u organismo.}
 
 \item{lat}{vector numérico con las latitudes límite de la región de interés.}
 
@@ -29,7 +29,7 @@ longitud) y el nombre.
 }
 \details{
 Esta función por defecto devuelve la lista completa de estaciones pero alternativamente
-se puede devolver estaciones especificar a partir de sus códigos o todas las estaciones
+se puede devolver estaciones específicas a partir de sus códigos o todas las estaciones
 incluidas en una región. Además incluye un método plot para visualizar rápidamente
 la ubicación de las estaciones.
 }
