+++ R/NH0358.R
@@ -1,3 +1,17 @@
-#' Datos diarios de estación
+#' Observaciones diarias de estaciones meteorológicas.
 #'
+#' Los datos provienen de 2 estaciones del INTA. Ver [metadatos_nh()] para más
+#' información y metadatos de las estaciones.
+#'
+#' @seealso [metadatos_nh()] devuelve los metadatos de las estaciones meteorológicas.
+#'
+#' @name datos_nh
+#' @rdname datos_nh
+#' @aliases NH0358,NH0114
+#' @format NH0358 es un data.table con 25515 filas y 25 columnas
 "NH0358"
+#'
+#' @name datos_nh
+#' @rdname datos_nh
+#' @format NH0114 es un data.table con 24041 filas y 25 columnas
+"NH0114"
