### 2024-05-27 20e5823
Mejoro la documentación
Close #42
