### 2024-05-27 5feaf9a
Prepara datos
