### 2024-05-27 1babf75
Arregla argumento resolucion sin numero #44
