### 2024-05-27 0293fb3
Comenta comportamiento de rango
