### 2024-05-28 5468363
Mejora documentacion Close #43
