### 2024-05-28 366210e
Mejora documentacion Close #44
