### 2024-05-28 4592a3f
Mejora documentación asociada a las funciones de mapear y ejemplos Close #47
