### 2024-05-28 6743a96
Mejora documentación SPEI
Close #52
