+++ R/spi.R
@@ -1,14 +1,23 @@
 #' Calcula el SPI
 #'
-#' Calcula esl Índice Estandarizado de Precipitación para distintas escalas. El `spi_indice` toma valores
-#' de precipitación mientras que el `spei_indice` toma valores del balance entre precipitación y evapotranspiración
-#' potencial. Internamente hacen lo mismo; la única diferencia es la distribución teórica
+#' Calcula esl Índice Estandarizado de Precipitación para distintas escalas. El
+#' `spi_indice` toma valores de precipitación mientras que el `spei_indice` toma
+#' valores del balance entre precipitación y evapotranspiración potencial.
+#' Internamente hacen lo mismo; la única diferencia es la distribución teórica
 #' usada por defecto para ajustar los datos.
+#'
+#' @details
+#' Estas funciones usan internamente a la función [SPEI::spi] pero tienen la ventaja
+#' de devolver el resultado como un data.frame que se puede usar de manera directa
+#' para el análisis de datos con dplyr.
+#'
 #'
 #' @param fecha vector de fechas.
 #' @param precipitacion vector de precipitacion.
 #' @param balance balance entre precipitación y evapotaranspiración potencial.
-#' @param escalas vector numérico con las escalas requeridas
+#' @param escalas vector numérico con las escalas requeridas. La unidad de la escala
+#' está dada por el vector de fechas. Si `escalas = 6` y los datos son mensuales
+#' entonces el cálculo del indice se hará en escalas de 6 meses.
 #' @param distribucion distribución usada para ajustar los datos.
 #' @param referencia serie de precipitación para usar de referencia en el ajuste
 #' a la distribución teórica. Puede ser:
@@ -24,6 +33,13 @@
 #' * `escala` (numérico) definidas en el argumento de entrada
 #' * `spi` o `spei` (numérico)
 #'
+#' @references
+#' Vicente-Serrano, S. M., Beguería, S. and López-Moreno, J. I.: A multiscalar
+#' drought index sensitive to global warming: The standardized precipitation
+#' evapotranspiration index, J. Clim., 23(7), \doi{10.1175/2009JCLI2909.1}, 2010.
+#'
+#' R Package [SPEI: Calculation of the Standardized Precipitation-Evapotranspiration Index](https://cran.r-project.org/package=SPEI)
+##
 #' @examples
 #' # datos aleatorios
 #' datos <- data.frame(fecha = seq(as.Date("1985-01-01"), as.Date("2015-12-01"), by = "1 month"))
