### 2024-05-29 33bf591
Actualizo ejemplo con la nueva funcionalidad de mapear() #47
