### 2024-05-29 15d1c52
Actualizo test de mapear() #47
