+++ tests/testthat/test-escalas.R
@@ -9,7 +9,6 @@ test_that("escalas", {
   escala1 <- scale_color_inta(escala = escala_pp_diaria$paleta)
   escala2 <- ggplot2::discrete_scale("color",
                            name = waiver(),
-                           scale_name = "precipitacion",
                            palette = escala_pp_diaria$paleta,
                            breaks = waiver(),
                            drop = TRUE)
