### 2024-05-29 f95f067
Actualizo ejemplos de mapear() en otras fucniones #47
