### 2024-05-29 6c6d2c3
Mejora documentación Close #50
