+++ man/umbrales.Rd
@@ -17,7 +17,8 @@ Variables fijas
 \itemize{
 \item \code{extremo} (caracter) nombre del extremo definido por el usuario
 (si los argumentos de \code{...} no tienen nombre, se usa \code{V1}, \code{V2}, etc...)
-\item \verb{N`` (numérico) ocurrencia del evento *}prop` (numérico) proporción eventos respecto del total de observaciones
+\item \code{N} (numérico) ocurrencia del evento
+*\code{prop} (numérico) proporción eventos respecto del total de observaciones
 \item \code{na} (numérico)  proporción de datos faltantes respecto del total de observaciones
 }
 }
@@ -32,6 +33,7 @@ Esto permite calcular distintos umbrales y obtener resultados para distintos gru
 \examples{
 data(NH0358)
 library(dplyr)
+
 # Sin agrupar devuelve un único valor
 NH0358 \%>\%
  summarise(umbrales(t_30 = t_max >= 30))
