### 2024-05-29 90de500
Arregla documentación y separa coeficientes
#51
