+++ R/pdsi.R
@@ -3,24 +3,32 @@
 #' Calcula el Indice de Severidad de Sequía de Palmer.
 #' `psdi_ac()` calcula la versión autocalibrada.
 #'
-#' @param precipitacion serie de precipitación sin datos faltantes.
+#' @param precipitacion serie de precipitación sin datos faltantes (en mm).
 #' @param etp serie de evapotranspiración potencial sin datos faltantes.
 #' @param cc capacidad de campo (en mm).
-#' @param coeficientes list de coeficientes que devuelve `pdsi_coeficientes()`
+#' @param coeficientes lista de coeficientes que devuelve `pdsi_coeficientes()`
 #'
 #' @return
 #' Un vector de la misma longitud que `precipitacion` con el PSDI correspondiente a cada caso.
 #'
 #' @details
 #'
-#' El Índice de Severidad de Sequía de Palmer, propuesto por Palmer (1965) es usado como indicador para
-#' cuantificar las condiciones de sequía a largo plazo. El el cálculo usa constantes definidas empíricamente,
-#' originalmente utilizando datos meteorológicos de Kansas y de Iowa en Estados Unidos. Estas constantes no
-#' representan necesariamente cualquier región del planeta por lo que puede ser redefinidas para el cálculo
-#' del índice usando la función `pdsi_coeficientes()`.
+#' El Índice de Severidad de Sequía de Palmer, propuesto por Palmer (1965) es usado
+#' como indicador para cuantificar las condiciones de sequía a largo plazo. El
+#' cálculo usa constantes definidas empíricamente, originalmente utilizando datos
+#' meteorológicos de Kansas y de Iowa en Estados Unidos. Estas constantes no
+#' representan necesariamente cualquier región del planeta por lo que puede ser
+#' redefinidas para el cálculo del índice usando la función
+#' del índice usando la función [pdsi_coeficientes()].
 #'
 #' Alternativamente, Wells et al. (2004) propuso el Indice de Severidad de Sequía de Palmer Autocalibrado
 #' que tiene la capacidad de ajustar las constantes empíricas durante el cálculo del indice.
+#'
+#' @references
+#' Palmer (1965), Meteorological Drought. U.S Weather Bureau, Washington, D.C. (book).
+#'
+#' Wells et. al. (2004), A Self-Calibrating Palmer Drought Severity Index. Journal
+#' of Climate \doi{10.1175/1520-0442(2004)017<2335:ASPDSI>2.0.CO;2}
 #'
 #' @examples
 #' # datos aleatorios
@@ -51,21 +59,4 @@ pdsi_ac  <- function(precipitacion, etp, cc = 100, coeficientes = pdsi_coeficien
   etp[missing] <- NA
 
   as.vector(pdsi_internal(precipitacion, etp, AWC = cc, sc = TRUE)$X)[seq_along(precipitacion)]
-}
-
-#' @rdname pdsi
-#'
-#' @param p,q factores de duración
-#' @param K1.1,K1.2,K1.3,K2 coeficientes de características climáticas
-#' @export
-
-pdsi_coeficientes <- function(p = 0.897,
-                              q = 1/3,
-                              K1.1 = 1.5,
-                              K1.2 = 2.8,
-                              K1.3 = 0.5,
-                              K2 = 17.67) {
-
-  list(p = p, q = q, K1.1 = K1.1, K1.2 = K1.2, K1.3 = K1.3, K2 = K2)
-
 }
