### 2024-06-01 d2d9447
Arregla errores en viñeta close #53
