+++ vignettes/estadisticas-e-indices-climaticos.Rmd
@@ -2,11 +2,11 @@
 title: "Estadísticas e índices climáticos"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Estadísticas e índices climáticos}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
+ %\VignetteIndexEntry{Estadísticas e índices climáticos}
+ %\VignetteEngine{knitr::rmarkdown}
+ %\VignetteEncoding{UTF-8}
 ---
-
+  
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
 collapse = TRUE,
@@ -60,13 +60,15 @@ plot(metadatos_nh(lon = c(-75, -71)))
 
 ## Análisis de la precipitación
 
-
+Trabajaremos con los datos de ejemplo provistos por el paquete, en este caso 
+de la estación meteorológica Castelar (NH0358). Para eso calculamos los valores
+mensuales.
 
 ```{r}
 datos_mensuales <- NH0358 %>% 
   group_by(fecha = lubridate::round_date(fecha, "month")) %>% 
   reframe(precip = mean(precip, na.rm = TRUE),
-            etp = mean(etp, na.rm = TRUE))
+          etp = mean(etp, na.rm = TRUE))
 
 head(datos_mensuales)
 ```
@@ -87,7 +89,7 @@ head(datos_mensuales)
 
 Valores cercanos a cero implican que la precipitación de ese mes fue similar a su valor promedio. 1 indica que llovió el doble de lo normal, mientras que -0.5 significa que en ese mes llovió la mitad de lo que suele llover. 
 
-Si la idea es usar esta medición para monitoreo, es importante fijar el período de referencia sobre el cual se calcula la precipitación media. De otra forma, a medida que se recolectan más datos, los promedios van a variar y con ellos los valores calculados. Entonces, para asegurarse de que los datos futuros no modifiquen los percentiles pasados, se puede especificar el período de referencia con el argumento `referencia`. Por ejemplo, este código devuelve el desvío de cada mes con respecto a la media anterior a 1980. 
+Si la idea es usar esta medición para monitoreo, es importante fijar el período de referencia sobre el cual se calcula la precipitación media (o climatología). De otra forma, a medida que se recolectan más datos, los promedios van a variar y con ellos los valores calculados. Entonces, para asegurarse de que los datos futuros no modifiquen los percentiles pasados, se puede especificar el período de referencia con el argumento `referencia`. Por ejemplo, este código devuelve la anomalía porcentual de cada mes con respecto a la media anterior a 1980. 
 
 ```{r}
 datos_mensuales <- datos_mensuales %>% 
@@ -96,7 +98,7 @@ datos_mensuales <- datos_mensuales %>%
   ungroup()
 ```
 
-`referencia` también puede ser un vector numérico de precipitación. Esto es útil si se calcula el valor de referencia a parte y luego sólo se leen los nuevos datos. 
+`referencia` también puede ser un vector numérico de precipitación. Esto es útil si se calcula el valor de referencia aparte y luego sólo se leen los nuevos datos. 
 
 Otras funciones de agroclimatico tienen este argumento, así que para mantener este período fijo, se puede crear una nueva columna. 
 
@@ -109,7 +111,7 @@ head(datos_mensuales)
 
 ### Deciles
 
-Otor indicador que puede analizarse es el decil al que pertenece la precipitación de cada mes. 
+Otro indicador que puede analizarse es el decil al que pertenece la precipitación de cada mes. 
 
 ```{r}
 datos_mensuales <- datos_mensuales %>% 
[…]
