### 2024-06-01 af51355
Caracteres no ASCCI en datos
