### 2024-06-01 8ae5fd4
Separa coeficientes de pdsi Close #51
