+++ R/pdsi_coeficientes.R
@@ -24,6 +24,18 @@
 #' Wells et. al. (2004), A Self-Calibrating Palmer Drought Severity Index. Journal
 #' of Climate \doi{10.1175/1520-0442(2004)017<2335:ASPDSI>2.0.CO;2}
 #'
+#' @examples
+#'
+#' # datos aleatorios
+#' datos <- data.frame(fecha = seq(as.Date("1985-01-01"), as.Date("2015-12-01"), by = "1 month"))
+#' set.seed(42)
+#' datos$pp <- rgamma(nrow(datos), shape = 2, scale = 10)
+#' datos$etp <- rgamma(nrow(datos), shape = 1, scale = 3)
+#'
+#' datos$pdsi_ac <- with(datos, pdsi(pp, etp),
+#'    coeficientes = pdsi_coeficientes(p = 0.9, q = 1/3))
+#'
+#'
 #' @export
 
 pdsi_coeficientes <- function(p = 0.897,
