+++ man/decil.Rd
@@ -30,7 +30,7 @@ de la anomalía porcentual.}
 \value{
 Devuelve un vector numérico con el decil asociado a cada valor de la variable.
 En este caso la columna \code{deciles} es de tipo doble ya que devuelve el valor exacto
-del decil sin redondoes. Para el cálculo de la anomalía porcentual también
+del decil sin redondeos. Para el cálculo de la anomalía porcentual también
 devuelve un vector numérico. Las funciones son compatibles con
 \code{\link[dplyr:group_by]{dplyr::group_by()}} y \code{\link[dplyr:mutate]{dplyr::mutate()}}.
 }
