### 2024-06-01 1b21f35
Arregla ejemplos largos
