+++ R/ith.R
@@ -21,7 +21,8 @@
 #' * Armstrong, DV. 1994. Heat stress interaction with shade and cooling. J. Diary Sci. 77:2004-2050
 #'
 #' @examples
-#' ith(temperatura = 23, hr = 65)
+#' ith(temperatura = 23, hr = 65) %>%
+#' slice_head(n = 10)
 #'
 #' data(NH0358)
 #'
@@ -29,7 +30,8 @@
 #' library(dplyr)
 #' NH0358 %>%
 #'   mutate(t_media = (t_max + t_min)/2) %>%
-#'   mutate(ith = ith(t_media, hr))
+#'   mutate(ith = ith(t_media, hr)) %>%
+#' slice_head(n = 10)
 #'
 #' @export
 ith <- function(temperatura, hr) {
