+++ R/olas.R
@@ -27,7 +27,8 @@
 #'
 #' library(dplyr)
 #' NH0358 %>%
-#'   summarise(olas(fecha, calor = t_max > 20, frio = t_min <= 0))
+#'   reframe(olas(fecha, calor = t_max > 20, frio = t_min <= 0)) %>%
+#'   slice_head(n = 10)
 #'
 #' @export
 olas <- function(fecha, ...) {
