+++ man/completar_serie.Rd
@@ -41,6 +41,7 @@ datos_perdidos <- sample(nrow(datos), nrow(datos)/10)
 datos$pp[datos_perdidos] <- NA
 datos_incompletos <- na.omit(datos) #Serie de datos a completar
 
-completar_serie(datos_incompletos, fechas, resolucion = "1 mes")
+completar_serie(datos_incompletos, fechas, resolucion = "1 mes") \%>\%
+slice_head(n = 10)
 
 }
