+++ man/olas.Rd
@@ -40,6 +40,7 @@ data(NH0358)
 
 library(dplyr)
 NH0358 \%>\%
-  summarise(olas(fecha, calor = t_max > 20, frio = t_min <= 0))
+  reframe(olas(fecha, calor = t_max > 20, frio = t_min <= 0)) \%>\%
+  slice_head(n = 10)
 
 }
