+++ .github/workflows/pkgcheck.yaml
@@ -17,5 +17,7 @@ on:
 jobs:
   pkgcheck:
     runs-on: ubuntu-latest
+    permissions:
+      issues: write
     steps:
       - uses: ropensci-review-tools/pkgcheck-action@main
