+++ man/spi_indice.Rd
@@ -64,6 +64,9 @@ de devolver el resultado como un data.frame que se puede usar de manera directa
 para el análisis de datos con dplyr.
 }
 \examples{
+
+library(dplyr)
+
 # datos aleatorios
 datos <- data.frame(fecha = seq(as.Date("1985-01-01"), as.Date("2015-12-01"), by = "1 month"))
 set.seed(42)
