### 2024-06-01 cb39298
Mejora return de metadatos #46
