+++ R/escalas.R
@@ -2,29 +2,16 @@
 #'
 #' Escalas de colores típicas usadas por INTA para distintas variables.
 #'
-#' @format
-#' Listas en el mismo formato que devuelve `leer_surfer()` con elementos:
+#' @format Objeto tipo lista con 3 elementos.
+#'
+#' @return
+#' Lista en el mismo formato que devuelve `leer_surfer()` con elementos:
 #' * `niveles` (numérico), el nivel a que corresopnde cada color.
 #' * `colores` (caracter), la representación hexadecimal del color de cada break.
 #' * `paleta` (función), una función que toma un entero `n` y devuelve
 #' un vector de caracter con `n` colores interpolados a partir de los colores
 #' de la escala.
 #'
-#' @rdname escalas
-"escala_temp_min"
-
-#' @format NULL
-#' @rdname escalas
-"escala_temp_max"
-
-#' @format NULL
-#' @rdname escalas
-"escala_pp_mensual"
-
-#' @format NULL
-#' @rdname escalas
-"escala_pp_diaria"
-
 #' @examples
 #'
 #' library(ggplot2)
@@ -45,5 +32,19 @@
 #'  geom_contour(aes(z = var1.pred)) +
 #'  geom_contour_filled(aes(z = var1.pred)) +
 #'  scale_fill_inta(escala = escala_pp_mensual)
+#'
+#' @rdname escalas
+"escala_temp_min"
 
+#' @format NULL
+#' @rdname escalas
+"escala_temp_max"
+
+#' @format NULL
+#' @rdname escalas
+"escala_pp_mensual"
+
+#' @format NULL
+#' @rdname escalas
+"escala_pp_diaria"
 
