### 2024-06-01 #599 @paocorrales
Hola!

En primer lugar, muchas gracias @pmnatural y @VeruGHub por la revisión y los comentarios, y a @Pakillo por ordenarlos y organizar todo.

Incorporé los comentarios y sugerencias al paquete. El detalle fino de los cambios se puede seguir en los [issues del repositorio del paquete](https://github.com/AgRoMeteorologiaINTA/agroclimatico/issues?q=is%3Aissue+is%3Aclosed+label%3A%22rOpenSci+review%22) (1 issue por función o familia de funciones).

Respondiendo a los comentarios específicos:

> #### dias_promedio
> No calcula ningun estadístico meteorológico como dice la descripción si no que extrae dia, mes y dia juliano de un set de datos. Previamente el usuario puede haber extraido los días que representan un evento meteorológico de interés, pero no tiene por qué.
>
> La descripción de la funcionalidad no es clara en la ayuda. Sugiero poner "Calcula el valor promedio anual del primer y ultimo dia para una serie de fechas dada". En los ejemplos habría que cambiar "summarise" por "reframe" como aparece en el Readme. En "Value" se dice que el output tiene "variables extras en el caso de hacer el cálculo para distintos grupos", pero ¿cómo se puede especificar para qué grupos quieres el cálculo?
>
> Erratas en la documentación: "seria de fechas"
>
> ```
> dias_promedio(datos$fecha) #Ok
> dias_promedio(NH0358$fecha) #Ok
>
> helada <- NH0358 |>
>   filter(t_min <= 0)
>
> dias_promedio(helada$fecha) #Ok
>
> helada |> summarise(dias_promedio(fecha)) #En un pipe, solo si la funcion se escribe dentro de un summarise o reframe se consigue la funcionalidad descrita
>
> #La función no devuelve un output que encaje bien con la filosofía de tidyverse y es difícil de utilizar en un pipe:
>
> helada |> dias_promedio(fechas = fecha)
> helada |> mutate(diasprom = dias_promedio(fecha))
> ```
Mejoré la documentación para explicar mejor el uso de la función junto con los verbos de dplyr. Corregí errores en los ejemplos (ahora usa reframe()). Incorporé un nuevo ejemplo que muestra como usar la función agrupando datos. Para esto el paquete tiene un nuevo set de datos que corresponde a una serie temporal de una estación meteorológica similar al que ya se estaba usando.

> #### completar_serie
> No se especifica el formato del vector fecha ni de rango en la documentación.
>
> ```
> > #Ejecuto el ejemplo #Ok
> ejemplo <- data.frame(fechas = seq(as.Date("1985-01-01"), as.Date("2015-12-01"),
>                              by = "1 month"),
>                     pp = 1)
>
> set.seed(42)
> datos_perdidos <- sample(nrow(ejemplo), nrow(ejemplo)/10)
> ejemplo$pp[datos_perdidos] <- NA
> datos_incompletos <- na.omit(ejemplo)
>
> completar_serie(datos_incompletos, fechas, resolucion = "1 mes")
>
> #No funciona, pero sigue las instrucciones de la documentación:
> completar_serie(datos = datos,
>                 fecha = fecha,
>                 resolucion = "día",
>                 rango = as.Date(c("1951-01-01", "1952-12-31")))
> ```
Había un error en el código. Ahora funciona tanto si la resolución es "dia" o "1 dia". Mejoré la documentación agregando los detalles a los argumentos mencionados.

> #### decil y anomalia_porcentual
> En la descripción se necesita incluir una definición más técnica de qué valores se devuelven, es decir, que hace la función decil y la anomalia_porcentual y referencias. En estadística, los deciles de una serie de datos son estadísticos descriptivos que representan los valores que limitan la serie de datos en 10 partes. Al aplicar la función sin tener mucho detalle en la descripción yo hubiera esperado obtener eso. Sin embargo, en el ejemplo ejecutado debajo obtengo un vector de igual longitud de la serie de datos. Sugiero mejorar la descripción en este sentido. La anomalía porcentual ¿respecto a qué valor de la serie de referencia se calcula? Debería especificarse. ¿Por qué para calcular la anomalia es necesario especificar que no se consideren los NA, pero para decil no? Parece una cuestión de inconsistencia en cómo están escritas las funciones.
>
> ```
> decil(datos$t_max) #Ok
>
> anomalia_porcentual(datos$t_max, na.rm = TRUE) #Ok
> ```
Mejore la documentación para explicar mejor la salida de la función. Respecto a la diferencia entre las funciones, efectivamente ambas son internamente distintas y deciles() no requiere el argumento na.rm. Las documentamos en conjunto ya que se suelen usar en conjunto y encontramos muchos ejemplos donde funciones documentadas como "familia" comparten algunos pero no todos los argumentos documentados.

> #### scala_temp/pp_XXX
> La documentación está incompleta. No hay ejemplos de uso ni se entiende bien que hacen estas funciones. Tampoco se  especifica qué valores retorna.

Agregue ejemplos en la documentación y lo que devuelve. Los ejemplos usan los nuevos datos del paquete y son mas simples.

> #### ith
> Funciona correctamente y está bien documentada.
>
> ```
> NH0358 <- NH0358 %>% #Ok
>   mutate(t_media = (t_max + t_min)/2) %>%
>   mutate(ith = ith(t_media, hr))
>
> hist(NH0358$ith)
> ```
Para esta y otras funciones revisé el output de los ejemplos y cuando eran muy largos ahora solo muestra las primeras 10 lineas.

> #### metadatos_nh
> Para facilitar la comprensión a los usuarios convendría especificar que son "estaciones NH". La descripción es incompleta en el sentido de que sólo se especifica una parte de la información que devuelve la función.
>
> Erratas en la documentación: "provincia o provincia", "especificar" debería ser "específicas".
>
> ```
>
> metadatos_nh(codigo = c("0001", "0011")) #Ok
> metadatos_nh(provincia ="Chaco") #Ok
> metadatos_nh(organismo = "SMN") #Ok
> metadatos_nh(lat = c(-30, -20), lon = c(-65, -55)) #Ok
> ```
Arreglé los errores de tipeo y extendí la documentación para incluir información sobre metadatos.

> #### kable_inta
> Sustituir "Setear" por "Formatear" o similar. Funciona correctamente.
>
> ```
>
> kable_inta(datos[1:10,1:5]) #Ok
> kable_inta(metadatos_nh(provincia ="Chaco")) #Ok
> ```
Documetación arreglada.

> #### leer_surfer
> En el ejemplo se sobreescribe un objeto con el nombre de otra función implementada en el paquete (scala_temp/pp_XXX). Debería estar explicado en algún sitio cual es el link entre ambas.
>
> Erratas en la documentación: "laeer".
>
> ```
>
> escala <- system.file("extdata", "escala_pp_mensual.lvl", package = "agroclimatico")
>
> escala_prueba <- leer_surfer(escala) #Ok
> scales::show_col(escala_prueba$colores)
> ```
Documentación arreglada y mejores ejemplos (asociadas a las funciones scale_*())

> #### mapa_XXX
> Añadiría en la descripción los posibles valores para el argumento "provincias".
>
> ```
>
> ggplot() +
>   geom_sf(data = mapa_argentina()) #Ok
>
> ggplot() +
>   geom_sf(data = mapa_provincias()) #Ok
>
> ggplot() +
>   geom_sf(data = mapa_provincias(provincias = "Chaco", departamentos =  TRUE)) #Ok
>
> ggplot() +
>   geom_sf(data = mapa_provincias(provincias = "Rio Negro", departamentos = FALSE)) #Ok
>
> ggplot() +
>   geom_sf(data = mapa_argentina_limitrofes()) #Ok
>
> ggplot() +
>   geom_sf(data = mapa_departamentos(provincias = "Chaco")) #Ok
> ```
Gracias por la sugerencia, ahora incluye la lista de provincias.

> #### mapear
> Erratas en la documentación: "veriable", "a el", "masyores", "1500m".
>
> El argumento "breaks" no está bien definido. Habría que especificar que se refiere a que hay que introducir valores (¿uno o varios?) en el rango del argumento "valor".
>
> No se explican las funciones coord_argentina ni theme_inta_mapa en ningún sitio. ¿Se utilizan sólo internamente? En ese caso no deberían estar exportadas.
>
> ```
>
> prueba <- metadatos_nh()
>
> mapear(prueba$altura, #Ok
>        prueba$lon,
>        prueba$lat,
>        breaks = c(0, 100, 400, 1000, 1500),
>        cordillera = 1400,
>        variable = "mm")
> ```
Arreglé los errores de tipeo e incluí las aclaraciones pedidas. Además ahora la función `mapear()` acepta el data.frame de datos como primer argumento. Las funciones secundarias  coord_argentina y theme_inta_mapa ahora tienen documentación.

> #### olas y umbrales
> No se explica en la documentación como hay que especificar los umbrales. Es poco habitual que se pueda poner cualquier palabra como nombre de un argumento y una mala praxis porque se puede sobreescribir algo importante en la función. Sería más adecuado que hubiera un argumento para los umbrales y que éstos se pudieran introducir en forma de lista con nombres (ej. umbral = list(calor = "t_max > 20"))
>
> Corregir algunas erratas en la documentación de olas: "mbral", "drtn"
>
> Corregir algunas erratas en la documentación de umbrales: "`` ⁠N`` (numérico) ocurrencia del evento *⁠ ``prop' "
>
> ```
> olas(datos$fecha, heatwave = datos$t_max > 20) #Ok
>
> umbrales(heatwave = datos$t_max > 20) #Ok
> ```
Corregí los errores de tipeo (drtn es un tipo de dato que representa diferencia entre tiempos). No cambié el comportamiento de las funciones `olas()` y `umbrales()`, siguen aceptando nombres de columnas como argumentos. Si bien entiendo el comentario de @VeruGHub, cuando desarrollamos estas funciones lo tuvimos en cuenta y decidimos mantener este comportamiento ya que hay muchas otras funciones de R que lo usan (por ejemplo data.frame() o summarise()). Por esto y teniendo en cuenta que las funciones no tienen otros argumentos definidos, consideramos que no hay riesgo de sobrescritura.

> #### pdsi/pdsi_ac
> No funciona indicandole los argumentos no definidos por defecto.
>
> Habría que incluir en la documentación las referencias que se citan e incluir las unidades de la precipitacion y la etp. Es redundante incluir un argumento para los coeficientes en forma de lista y luego argumentos para cada coeficiente por separado. Sugiero eliminar una de las dos formas.
>
> Corregir algunas erratas de la ayuda: "list de coeficientes", "El el cálculo".
> ```
>
> pdsi(precipitacion = c(500, 600), #No funciona
>      etp = c(1000, 1200))
>
> pdsi_ac(precipitacion = c(500, 600), #No funciona
>      etp = c(1000, 1200))
> ```
Extendí la documentación con todo lo mencionado. Además incluí la necesidad de tener una serie larga de datos para calcular el indice (no es suficiente con 2 datos). La función pdsi_coeficientes() ahora se documenta aparte para no generar confución.

> #### scale_color/fill_inta pdsi/pdsi_ac
> El ejemplo no funciona. Tampoco funciona otro ejemplo que he creado siguiendo las instrucciones de la documentación.
>
> Corregir erratas en la documentación: "escala.)", "tiene prioridad por sobre", "deinidos"
>
> ```
>
> prueba <- metadatos_nh()
> prueba$altura <- as.numeric(prueba$altura)
>
> ggplot(prueba) + #No funciona
>   geom_point(aes(x = lon, y = lat, color = altura)) +
>   scale_color_inta(escala = escala_pp_diaria)
>
>
> set.seed(496) #El ejemplo no funciona
> datos_aleatorios <- subset(metadatos_nh())
> datos_aleatorios <- data.frame(datos_aleatorios,
> pp = rgamma(nrow(datos_aleatorios), 0.5, scale = 1)*70)
>
> ggplot(datos_aleatorios, aes(lon, lat)) +
> geom_contour(aes(z = pp, fill = after_stat(level))) +
> scale_fill_inta(escala = escala_pp_diaria)
> ```
Mejoré la documentación de esta familia de funciones, en primer lugar indicando que son escalas de colores para variables discretas (no estaba explicito) y mejoré los ejemplos usando un nuevo set de datos con valores de precipitación mensual. Esto simplifica los ejemplos y son más realistas.

> #### spei_indice/spi_indice
> La funcionalidad está repetida respecto a las funciones spi y spei del paquete SPEI. A menos que se aclare qué aporta esta función sugiero eliminarla y crear una viñeta que incluya cómo se usa el paquete SPEI con datos en el formato del INTA. Si he entendido bien, las fechas deben introducirse como meses ¿no? Por favor, especificar esto si se mantiene la función. En la documentación no se entiende bien que es el argumento "escalas" y no se cita el paquete SPEI.
>
> ```
> prueba <- data.frame(fecha = seq(as.Date("1985-01-01"), as.Date("2015-12-01"), by = "1 month"))
> set.seed(42)
> prueba$pp <- rgamma(nrow(prueba), shape = 2, scale = 10)
>
> spi_indice(prueba$fecha, prueba$pp, escalas = 1)
>
>
> #Misma funcionalidad que en paquete SPEI:
> spi_agro <- spi_indice(fecha = prueba$fecha, precipitacion = prueba$pp, escalas = 1)
>
> spi_spei <- SPEI::spi(data = prueba$pp, scale = 1)
>
> identical(spi_agro$spi, as.vector(spi_spei$fitted))
> ```
Modifiqué la documentaición incluyendo lo solicitado. Una de las ventajas principales de estas funciones y que justifican su existencia es que devuelven el resultado como un data.frame que se puede usar de manera directa para el análisis de datos con dplyr.

Además de lo mencionado anteriormente:

* Apliqué los comentarios al README y actualicé lo necesario en base a los cambios en el código y nuevos datos. El logo ahora refleja el nuevo nombre del paquete.

* Actualicé la viñeta del paquete para que refleje los cambios hechos y aproveche los datos disponibles.

* El listado de funciones en la web del paquete ahora se ordena por temas.

* Hice una revisión de errores de tipeo en todo el paquete.

Espero no haberme olvidado de nada, aguardo sus comentarios!
