### 2024-06-12 #599 @Pakillo
Muchas gracias @paocorrales !

Estoy revisando todos los cambios. Me parece que la [página web](https://agrometeorologiainta.github.io/agroclimatico/index.html) del paquete no está actualizada a la última versión, ¿cierto? Veo que el GitHub Actions de [pkgdown](https://github.com/AgRoMeteorologiaINTA/agroclimatico/actions/workflows/pkgdown.yaml) está desactivado. Por favor actívelo o regenere la web con pkgdown manualmente. Nos ayudaría ver la web actualizada para la revisión

¡Gracias!
