### 2024-06-16 0e9e2c1
Fix typo in spi.R
