### 2024-06-16 1d38826
fix typo in mapear.R
