+++ R/mapear.R
@@ -1,7 +1,7 @@
 #' Grafica variables en Argentina
 #'
 #' Dadas mediciones de una variable en puntos ubicados en Argentina,
-#' interpola al resto del territorio usando kriging y grafica con contornos
+#' interpola al resto del territorio usando kriging y grafica con contornos.
 #' Las funciones secundarias [coord_argentina()] y [theme_inta_mapa()] permiten
 #' generar un mapa en la región de Argentina (definida por `xlim` y `ylim`) con
 #' el estilo específico usando por INTA.
