### 2024-06-16 #599 @Pakillo
¡Muchas gracias @paocorrales por la revisión! Creo que el paquete y la documentación han mejorado muchísimo. Algunos comentarios a la última versión:

## decil

Es posible que no esté entendiendo bien esta función, pero encuentro su valor de salida un poco confuso. Por ejemplo, en este caso, `decil` devuelve el decil (de 1 a 10) en que se encuentra cada valor concreto de la variable:

``` r
x = 1:10
decil(x)
```

    ##  [1]  1  2  3  4  5  6  7  8  9 10

Pero en este caso devuelve valores con decimales desde 0 hasta 10:

``` r
x <- seq(0, 19.5, 0.5)
decil(x)
```

    ## [37]  9.25  9.50  9.75 10.00

Pienso que estos valores con decimales podrían generar confusión. Por ejemplo, alguien que quiera saber en qué decil cae cada dato podría intentar hacer un `round` de estos valores, obteniendo esto:

``` r
round(decil(x))
```

    ## [26]  6  7  7  7  8  8  8  8  8  9  9  9 10 10 10

cuando probablemente necesite ejecutar

``` r
ceiling(decil(x))
```

    ## [26]  7  7  7  8  8  8  8  9  9  9  9 10 10 10 10

¿Cuál es la ventaja de devolver números decimales como salida de `decil`? ¿Pienso que tal vez sería deseable devolver el decil como número entero comprendido entre 1 y 10? Creo que @VeruGHub apuntaba hacia la misma sugerencia en su revisión.

@VeruGHub también planteaba esta cuestión: ¿Por qué para calcular la anomalia es necesario especificar que no se consideren los NA, pero para decil no? Entiendo que la función `decil` ignora los NA por defecto, pero tal vez eso debería ser una opción que eligiera el usuario (esto es, añadir un argumento `na.rm`).

Por ejemplo, la función `quantile` da un error si le pedimos los deciles de un vector con NA:

``` r
x <- 1:10
x[1] <- NA
quantile(x, probs = seq(0, 1, 0.1))
```

    ## Error in quantile.default(x, probs = seq(0, 1, 0.1)): missing values and NaN's not allowed if 'na.rm' is FALSE

Si especificamos `na.rm = TRUE` entonces sí funciona:

``` r
quantile(x, probs = seq(0, 1, 0.1), na.rm = TRUE)
```

    ##  2.0  2.8  3.6  4.4  5.2  6.0  6.8  7.6  8.4  9.2 10.0

Al igual que @VeruGHub, pienso que la función `decil` podría/debería incorporar este argumento por consistencia con `quantile` y `anomalia_porcentual` y para mayor control de los usuarios. Si no, la función `decil` siempre arroja un resultado que podría ser poco robusto si la variable contiene muchos NA.

## dias_promedio

Esta función ha quedado muy bien, y la documentación y los ejemplos son muy claros. ¡Gracias!

## umbrales

El último ejemplo requiere cambiar de `summarise` a `reframe`.

## olas

Creo que sería bueno incluir un argumento `na.rm` o similar para evitar inconsistencias y que el usuario pueda elegir qué hacer con los datos faltantes.

Por ejemplo, esta ola de calor de 22 días de duración:

``` r
data(NH0358)
NH0358 %>%
  reframe(olas(fecha, calor = t_max > 20)) %>%
  slice_head(n = 1)
```

    ## 1 calor 1951-01-01 1951-01-22  22 days

se transforma en otra de 19 días de duración si falta el dato de temperatura del 20 de enero:

``` r
NH0358 %>%
  mutate(t_max = if_else(fecha == "1951-01-20", NA, t_max)) |>
  reframe(olas(fecha, calor = t_max > 20)) %>%
  slice_head(n = 1)
```

    ## 1 calor 1951-01-01 1951-01-19  19 days

pero vuelve a tener 22 días de duración si falta la fila completa del 20 de enero (cuando en realidad ese día la temperatura pudo ser \<20ºC):

``` r
NH0358 %>%
  filter(fecha != "1951-01-20") |>
  reframe(olas(fecha, calor = t_max > 20)) %>%
  slice_head(n = 1)
```

    ## 1 calor 1951-01-01 1951-01-22  22 days

Pienso que para evitar posibles inconsistencias, podría añadirse un nuevo argumento `na.rm` o similar, de manera que los usuarios puedan elegir qué hacer con los datos faltantes, y que el resultado sea consistente. Si el usuario quiere que los NA se tengan en cuenta, pienso que la función debería devolver NA como fecha de inicio, fin, o duración, cuando sea imposible determinar el día concreto de inicio o fin (por haber datos faltantes).

Por otro lado, es una decisión de los autores, pero como sugerencia, ¿tal vez utilizar “duración” en lugar de “longitud” para especificar la duración (en días) de la ola?

## ith

El ejemplo devuelve todos los valores de ith como NA, dado que `hr` es NA en las 10 primeras filas. Tal vez podemos filtrar primero:

``` r
NH0358 %>%
  filter(!is.na(hr)) |>
  mutate(t_media = (t_max + t_min)/2) %>%
  mutate(ith = ith(t_media, hr)) %>%
  slice_head(n = 10)
```

## PDSI

El ejemplo usa `with` y `$` para crear variables, pero en coherencia con la filosofía ‘tidy’ del paquete ¿tal vez sea preferible utilizar código más tipico del tidyverse?

``` r
datos <- data.frame(fecha = seq(as.Date("1985-01-01"),
                                as.Date("2015-12-01"),
                                by = "1 month"))
set.seed(42)
datos |>
  mutate(pp = rgamma(nrow(datos), shape = 2, scale = 10),
         etp = rgamma(nrow(datos), shape = 1, scale = 3),
         pdsi_ac = pdsi(pp, etp)) |>
  slice_head(n = 10)
```

    ## 10 1985-10-01  2.812335 4.3753527 -1.2851817

(modificar también el ejemplo de `pdsi_coeficientes`).

## SPI

Creo que sería bueno que los ejemplos de esta función siguieran el estilo utilizado en el [tutorial](https://agrometeorologiainta.github.io/agroclimatico/articles/estadisticas-e-indices-climaticos.html#%C3%ADndice-estandarizado-de-precipitaci%C3%B3n), esto es, usando `reframe` en lugar de `with`.

He corregido algunas erratas en este [pull request](https://github.com/AgRoMeteorologiaINTA/agroclimatico/pull/54).

Espero que mis comentarios/sugerencias sean claros y fáciles de solucionar. Quedo a la espera de estos cambios, ¡y ya casi estaríamos!

Gracias
