### 2024-07-29 94ed024
Agrega test para serie completa
