### 2024-07-29 f0bb3b3
Mejora ejemplos de pdsi
