+++ R/ith.R
@@ -28,6 +28,7 @@
 #' # En el contexto de mutate
 #' library(dplyr)
 #' NH0358 %>%
+#'   filter(!is.na(hr)) |>
 #'   mutate(t_media = (t_max + t_min)/2) %>%
 #'   mutate(ith = ith(t_media, hr)) %>%
 #' slice_head(n = 10)
