### 2024-07-29 0d53d25
Mejora ejemplos spi
