+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.1.0
 pkgdown_sha: ~
 articles:
   estadisticas-e-indices-climaticos: estadisticas-e-indices-climaticos.html
-last_built: 2024-07-29T06:42Z
+last_built: 2024-07-29T11:00Z
 urls:
   reference: https://agrometeorologiainta.github.io/agroclimatico/reference
   article: https://agrometeorologiainta.github.io/agroclimatico/articles
