### 2024-07-29 #599 @paocorrales
Hola @Pakillo! Gracias por los comentarios. Respondo a cada uno por separado. En el caso de olas(), la función incluye ahora los cambios pedidos y además se resuelve un error que surgió de los ejemplos. La función decil() no tiene cambios aún pero con @eliocamp damos más detalles abajo y quedo atenta a sus respuestas.

## decil()

La función `decil()` lo que hace es una transformación en la que reemplaza cada valor por su cuantil asociado en base a la densidad de probabilidad acumulada de los datos.
Es decir, no es una función que discretice los datos en intervalos. Esto es así porque la distribución de una variables es continua.
Se llama decil porque el resultado se multiplica por 10 para que el máximo sea 10 en vez de 1.
(`decil(1:10)` devuelve valores redondos porque justo los datos de entrada coinciden exactamente con los límites de los deciles).

Este tipo de transformación es útil en ciertas comparaciones entre observaciones y modelos o entre múltiples modelos, ya que muchas veces los modelos tienen algún sesgo en la media o la varianza pero aún así representan bien la variabilidad relativa.

Cambiar el valor de retorno a valores redondeados implicaría cambiar la función de una transformación a una discretización, perdiendo información.
La discretización se puede realizar luego con el nivel de precisión que le usuarie requiera según su criterio (cantidad de datos, facilidad de visualización, consistencia con otros métodos, etc.)

```r
decil_x <- decil(rnorm(100))
cut(decil_x, seq(0, 10, by = 1))    # discretizar a deciles
cut(decil_x, seq(0, 10, by = 0.1))  # discretizar a percentiles
cut(decil_x, seq(0, 10, by = 2))    # discretizar a quintiles
```

La función no requiere trato especial de los valores faltantes porque al estimar la curva de densidad de probabilidad acumulado, éstos se descartan automáticamente (ver `?ecdf`).
Otras metodologías para estimar la densidad de probabilidad acumulada sí podrían ser sensibles a los datos faltantes (por ejemplo, se podría asumir que la distribución es normal y estimar sus parámetros con `mean()` y `sd()`).

También se podría devolver todo `NA` si algún valor es `NA` y agregar un argumento `na.rm` para controlar ese comportamiento.
No sé si eso solucionaría el problema de que la estimación de la CDF sea poco robusta, ya que fallaría aún en caso de que haya un único `NA` que no afectaría significativamente la estimación, y quien la use podría usar el `na.rm = TRUE` y obtener un resultado sin tener en cuenta la cantidad de datos que tiene.

## umbrales()

Ejemplo arreglado

## olas()

El ejemplo que presentas es muy interesante porque detecta un problema que no habíamos tenido en cuenta. La inconsistencia que detectas no se debe a la presencia de NAs explicitos, si no que se debe a un dato faltante implicito, la falta de un día dentro de la serie de datos.
No habiamos tenido en cuenta esta situación y al calcular la diferencia entre las fechas de fin y de inicio obtenemos el mismo resultado aún si faltan dias en el medio.

Agregamos un test dentro de la función que chequea que la serie que recibe está completa. Si no está completa devuelve un error.

Respecto al comportamiento para NAs, sumamos un argumento remplaza.na que por defecto sigue devolviendo el mismo resultado. Esto es consistente con la función rle() de base. Sin embargo si es TRUE, la función reemplazara los NA por el valor previo en la serie "sumando" ese valor a la ola o persistencia.

## ith()

Gracias, ejemplo arreglado.

## PDSI

Gracias por la recomendación!

## SPI

Estos ejemplos también están actualizados
