### 2024-08-07 #599 @Pakillo
¡Muchas gracias por la respuesta y los cambios implementados! Me alegro de que todo haya tenido fácil solución.

Gracias también por explicar lo que hace la función `decil`. Ahora lo entiendo (creo): para cada valor, devuelve el cuantil basado en una función de probabilidad acumulada (pero en lugar de devolver directamente el cuantil, multiplica el valor por 10). Claramente no había entendido bien su finalidad (y pienso que @VeruGHub tal vez tampoco, a juzgar por sus comentarios). Ello me lleva a proponer un cambio de nombre de esta función, pues pienso que al igual que a nosotros, puede confundir a otras personas en el futuro, que piensen que la función `decil` les va a devolver los deciles (cuando no es así). No sé si `cuantil` podría ser un nombre más adecuado, menos propenso a confusión. También me pregunto sobre la conveniencia de devolver los valores multiplicados por 10, en lugar de usar una escala de 0 a 1? Por otro lado, pienso que sería bueno mejorar la ayuda de la función con las explicaciones dadas más arriba: "reemplaza cada valor por su cuantil asociado en base a la densidad de probabilidad acumulada de los datos". Esta misma explicación sugiere que "cuantil" podría ser un buen nombre para esta función.

Tras esta propuesta para su consideración parte de los autores, invito de nuevo a @pmnatural y @VeruGHub a revisar el paquete y los cambios aquí discutidos. Por favor, comuniquen si tienen más observaciones que hacer o si no manifiesten su conformidad usando [esta plantilla](https://devguide.ropensci.org/approval2template.html).

¡Gracias!
