### 2024-09-20 #599 @Pakillo
Hola @pmnatural @VeruGHub

Cuando puedan por favor revisen si están conformes con los cambios realizados en el paquete y manifiesten su conformidad usando [esta plantilla](https://devguide.ropensci.org/approval2template.html).

Gracias
