### 2024-10-24 #599 @VeruGHub
Hola @paocorrales et al:

Muchas gracias por considerar los cambios propuestos. El readme y la documentación y ejemplos de las funciones han mejorado mucho y los cambios realizados han aclarado mis dudas. Disculpad @Pakillo y @paocorrales la tardanza de mi respuesta.

He detectado algunos puntos más que se pueden mejorar, por si se quieren considerar:

- En el readme sería mejor poner el enlace a la definición de datos tidy la primera vez que aparece este concepto.
- decil(): no habia entendido su funcionalidad como apunta @Pakillo. Coincido con él en que si el objetivo no es dar los deciles en que se encuentran los datos o los valores de la variable que marcan los límites de los deciles , un nombre mucho más adecuado sería "cuantil" o "cuantil10". La descripción del output que se da en "value", ya que hay dos funciones en la ayuda, se puede mejorar indicando claramente al inicio "Para el cálculo de los deciles, ...", al igual que luego se habla de las anomalías porcentuales. Estaría muy bien añadir un ejemplo para la función anomalia_porcentual().
- En la función olas() sugiero que el resultado de la duración esté en español como el resto de la salida (es decir, cambiar "days" por "dias" y lo mismo si hay otras unidades de medida en inglés. En el ejemplo, estaría bien que se vieran las olas de frio. Por ejemplo, se puede separar el cálculo de olas de calor y de frío en dos líneas de código:

```
data(NH0358)
library(dplyr)
NH0358 %>%
  reframe(olas(fecha, calor = t_max > 20)) %>%
  slice_head(n = 10)

NH0358 %>%
  reframe(olas(fecha, frio = t_min <= 0)) %>%
  slice_head(n = 10)
```
- pdsi()/pdsi_ac(): si el error que me salia se debe a que se necesita una serie de datos "larga" debería indicarse cuanto de largos tienen que ser los vectores de entrada o definir de alguna manera mejor estos argumentos.
- spi_indice()/spei_indice(): aunque se hace una alusión a la funcionalidad de ordenar en "details", añadiría en la descripción la funcionalidad completa, es decir, que calcula el SPI/SPEI y ordena los datos. Esta segunda parte es la principal novedad de la función y no aparece.

Eso es todo por mi parte, enhorabuena por el paquete.

Un saludo
