### 2024-11-14 #599 @Pakillo
Hola @paocorrales et al.

Creo que el paquete está prácticamente listo para ser aceptado. No sé qué opinan de las últimas sugerencias (renombrar decil por cuantil o similar, traducir days al español...). En cuanto me digan su opinión sobre estos últimos comentarios propondré la aceptación del paquete.

Quedo a la espera, muchas gracias
