### 2024-11-26 1979ead
spi devuelve tidy data!
