+++ R/spi.R
@@ -1,15 +1,15 @@
 #' Calcula el SPI
 #'
-#' Calcula el Índice Estandarizado de Precipitación para distintas escalas. El
-#' `spi_indice` toma valores de precipitación mientras que el `spei_indice` toma
+#' Calcula el Índice Estandarizado de Precipitación para distintas escalas. Las
+#' funciones `spi_indice` y `spei_indice` usan internamente a la función [SPEI::spi]
+#' pero tienen la ventaja de devolver el resultado como un data.frame que se
+#' puede usar de manera directa para el análisis de datos con dplyr.
+#'
+#' @details
+#' La función `spi_indice` toma valores de precipitación mientras que `spei_indice` toma
 #' valores del balance entre precipitación y evapotranspiración potencial.
 #' Internamente hacen lo mismo; la única diferencia es la distribución teórica
 #' usada por defecto para ajustar los datos.
-#'
-#' @details
-#' Estas funciones usan internamente a la función [SPEI::spi] pero tienen la ventaja
-#' de devolver el resultado como un data.frame que se puede usar de manera directa
-#' para el análisis de datos con dplyr.
 #'
 #'
 #' @param fecha vector de fechas.
