### 2024-11-26 9e46a5f
Aclaración sobre largo de la serie temporal
