### 2024-11-26 a1292d6
Cambia argumentos en theme() por actualización de ggplot
