### 2024-11-26 ae69aa8
Mejora doc y ejemplos
