+++ R/decil.R
@@ -12,9 +12,10 @@
 #' @param na.rm lógico. Define si se utilizan o no valores faltantes en el cálculo
 #' de la anomalía porcentual.
 #'
-#' @return Devuelve un vector numérico con el decil asociado a cada valor de la variable.
-#' En este caso la columna `deciles` es de tipo doble ya que devuelve el valor exacto
-#' del decil sin redondeos. Para el cálculo de la anomalía porcentual también
+#' @return Para el cálculo de los deciles, devuelve un vector numérico con el decil
+#' asociado a cada valor de la variable. En este caso la columna `deciles` es de
+#' tipo doble ya que devuelve el valor exacto del decil sin redondeos.
+#' Para el cálculo de la anomalía porcentual también
 #' devuelve un vector numérico. Las funciones son compatibles con
 #' [dplyr::group_by()] y [dplyr::mutate()].
 #'
@@ -40,9 +41,19 @@
 #' precip_mensual %>%
 #'   mutate(deciles = decil(precip))
 #'
-#' # Definiendo un periodo de referencia
+#' # Deciles definiendo un periodo de referencia
 #' precip_mensual %>%
 #'   mutate(deciles = decil(precip,
+#'                          referencia = lubridate::year(fecha) <= 1958))
+#'
+#' # Anomalia porcentual usando como referencia la serie completa
+#' NH0358 %>%
+#'   mutate(anomalia = anomalia_porcentual(precip)) %>%
+#'   slice_head(n = 10)
+#'
+#' #  Anomalia porcentual definiendo un periodo de referencia
+#' precip_mensual %>%
+#'   mutate(deciles = anomalia_porcentual(precip,
 #'                          referencia = lubridate::year(fecha) <= 1958))
 #'
 #' @export
