+++ DESCRIPTION
@@ -31,7 +31,7 @@ Encoding: UTF-8
 Language: es
 LazyData: true
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.3.1
+RoxygenNote: 7.3.2
 URL: https://github.com/AgRoMeteorologiaINTA/agroclimatico, https://agrometeorologiainta.github.io/agroclimatico/
 BugReports: https://github.com/AgRoMeteorologiaINTA/agroclimatico/issues
 Depends: 
