+++ pkgdown.yml
@@ -1,9 +1,9 @@
 pandoc: 3.1.11
-pkgdown: 2.1.0
+pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
   estadisticas-e-indices-climaticos: estadisticas-e-indices-climaticos.html
-last_built: 2024-07-29T12:36Z
+last_built: 2024-11-26T09:36Z
 urls:
   reference: https://agrometeorologiainta.github.io/agroclimatico/reference
   article: https://agrometeorologiainta.github.io/agroclimatico/articles
