### 2024-11-26 #599 @paocorrales
Hola y muchas gracias nuevamente por los comentarios. Voy respondiendo a cada uno:

* README: actualizado con la sugerencia.
* `decil()`: mejoré la descripción de la salida y sumé ejemplos para `anomalia_porcentual()`. Respecto del nombre, lo consulté y preferiríamos no cambiarlo ya que rompería código que actualmente es muy útilizado. En el área, el nombre de la función coincide con los conceptos típicos.  Por supuesto, si vemos que genera confusión para nuevos usuarios (que esperamos puedan aprovechar el paquete), podemos evaluar alternativas.
* `olas()`: actualicé la sección ejemplos. Respecto a la columna `duracion` en la salida, me parece una sugerencia super razonable. El problema es que esta columna es de tipo `difftime`, lo que permite luego usar paquetes como lubridate. Traducir la salida al español implicaría convertir la columna a caracter. Preferimos dejar la función como está con la esperanza de que en algún momento eso se traduzca desde R base y dejar del lado del usuario la traducción al español en caso de que tenga que presentar la tabla.
* `pdsi()`: agregamos la aclaración en los detalles de la documentación.
* `spi_indice()`: gracias por la sugerencias, aplicada a la documentación.
