### 2024-11-29 #599 @Pakillo
Estupendo, muchas gracias por las respuestas. Suena todo muy razonable :)

¡Pues yo creo que con esto podemos dar por concluido el proceso de revisión! Espero que haya sido útil. Enhorabuena y muchas gracias tanto a autoras como revisoras por participar en el proceso de revisión de software de rOpenSci.
