### 2024-11-29 #599 @ropensci-review-bot
Approved! Thanks @paocorrales for submitting and @VeruGHub, @pmnatural for your reviews! :grin:
