### 2024-12-04 #599 @VeruGHub
Hola @paocorrales, por mi parte de acuerdo en aparecer como revisora :) Enhorabuena otra vez por el paquete.
